Using exec to load function definitions
