--- src/refactored.py
@@ -1,5121 +1,0 @@
-import numpy as np
-
-def tcs_rates(t, y, kf, kr):
-	arr = np.zeros((29, 1))
-	arr[0, 0] = 500.0*y[3]*y[4]
-	arr[1, 0] = 100000000.0*y[0]*y[6]
-	arr[2, 0] = 0.01*y[5]*y[8]
-	arr[3, 0] = 170000.0*y[10]*y[4]
-	arr[4, 0] = 100000000.0*y[12]*y[9]
-	arr[5, 0] = 0.01*y[11]*y[8]
-	arr[6, 0] = 0.0001*y[14]*y[4]
-	arr[7, 0] = 100000000.0*y[12]*y[13]
-	arr[8, 0] = 170000.0*y[16]*y[4]
-	arr[9, 0] = 100000000.0*y[15]*y[18]
-	arr[10, 0] = 0.01*y[17]*y[8]
-	arr[11, 0] = 0.0001*y[20]*y[4]
-	arr[12, 0] = 100000000.0*y[18]*y[19]
-	arr[13, 0] = 170000.0*y[22]*y[4]
-	arr[14, 0] = 100000000.0*y[21]*y[24]
-	arr[15, 0] = 0.01*y[23]*y[8]
-	arr[16, 0] = 0.0001*y[26]*y[4]
-	arr[17, 0] = 100000000.0*y[24]*y[25]
-	arr[18, 0] = 170000.0*y[28]*y[4]
-	arr[19, 0] = 100000000.0*y[27]*y[30]
-	arr[20, 0] = 0.01*y[29]*y[8]
-	arr[21, 0] = 0.0001*y[32]*y[4]
-	arr[22, 0] = 100000000.0*y[30]*y[31]
-	arr[23, 0] = 500.0*y[34]*y[4]
-	arr[24, 0] = 100000000.0*y[33]*y[36]
-	arr[25, 0] = 0.01*y[35]*y[8]
-	arr[26, 0] = 500.0*y[38]*y[4]
-	arr[27, 0] = 100000000.0*y[37]*y[40]
-	arr[28, 0] = 0.01*y[39]*y[8]
-	return arr.reshape(-1)
-
-def tcs_rates_jac(t, y, kf, kr):
-	arr = np.zeros((29, 41))
-	arr[0, 0] = 0
-	arr[0, 1] = 0
-	arr[0, 2] = 0
-	arr[0, 3] = 500.0*y[4]
-	arr[0, 4] = 500.0*y[3]
-	arr[0, 5] = 0
-	arr[0, 6] = 0
-	arr[0, 7] = 0
-	arr[0, 8] = 0
-	arr[0, 9] = 0
-	arr[0, 10] = 0
-	arr[0, 11] = 0
-	arr[0, 12] = 0
-	arr[0, 13] = 0
-	arr[0, 14] = 0
-	arr[0, 15] = 0
-	arr[0, 16] = 0
-	arr[0, 17] = 0
-	arr[0, 18] = 0
-	arr[0, 19] = 0
-	arr[0, 20] = 0
-	arr[0, 21] = 0
-	arr[0, 22] = 0
-	arr[0, 23] = 0
-	arr[0, 24] = 0
-	arr[0, 25] = 0
-	arr[0, 26] = 0
-	arr[0, 27] = 0
-	arr[0, 28] = 0
-	arr[0, 29] = 0
-	arr[0, 30] = 0
-	arr[0, 31] = 0
-	arr[0, 32] = 0
-	arr[0, 33] = 0
-	arr[0, 34] = 0
-	arr[0, 35] = 0
-	arr[0, 36] = 0
-	arr[0, 37] = 0
-	arr[0, 38] = 0
-	arr[0, 39] = 0
-	arr[0, 40] = 0
-	arr[1, 0] = 100000000.0*y[6]
-	arr[1, 1] = 0
-	arr[1, 2] = 0
-	arr[1, 3] = 0
-	arr[1, 4] = 0
-	arr[1, 5] = 0
-	arr[1, 6] = 100000000.0*y[0]
-	arr[1, 7] = 0
-	arr[1, 8] = 0
-	arr[1, 9] = 0
-	arr[1, 10] = 0
-	arr[1, 11] = 0
-	arr[1, 12] = 0
-	arr[1, 13] = 0
-	arr[1, 14] = 0
-	arr[1, 15] = 0
-	arr[1, 16] = 0
-	arr[1, 17] = 0
-	arr[1, 18] = 0
-	arr[1, 19] = 0
-	arr[1, 20] = 0
-	arr[1, 21] = 0
-	arr[1, 22] = 0
-	arr[1, 23] = 0
-	arr[1, 24] = 0
-	arr[1, 25] = 0
-	arr[1, 26] = 0
-	arr[1, 27] = 0
-	arr[1, 28] = 0
-	arr[1, 29] = 0
-	arr[1, 30] = 0
-	arr[1, 31] = 0
-	arr[1, 32] = 0
-	arr[1, 33] = 0
-	arr[1, 34] = 0
-	arr[1, 35] = 0
-	arr[1, 36] = 0
-	arr[1, 37] = 0
-	arr[1, 38] = 0
-	arr[1, 39] = 0
-	arr[1, 40] = 0
-	arr[2, 0] = 0
-	arr[2, 1] = 0
-	arr[2, 2] = 0
-	arr[2, 3] = 0
-	arr[2, 4] = 0
-	arr[2, 5] = 0.01*y[8]
-	arr[2, 6] = 0
-	arr[2, 7] = 0
-	arr[2, 8] = 0.01*y[5]
-	arr[2, 9] = 0
-	arr[2, 10] = 0
-	arr[2, 11] = 0
-	arr[2, 12] = 0
-	arr[2, 13] = 0
-	arr[2, 14] = 0
-	arr[2, 15] = 0
-	arr[2, 16] = 0
-	arr[2, 17] = 0
-	arr[2, 18] = 0
-	arr[2, 19] = 0
-	arr[2, 20] = 0
-	arr[2, 21] = 0
-	arr[2, 22] = 0
-	arr[2, 23] = 0
-	arr[2, 24] = 0
-	arr[2, 25] = 0
-	arr[2, 26] = 0
-	arr[2, 27] = 0
-	arr[2, 28] = 0
-	arr[2, 29] = 0
-	arr[2, 30] = 0
-	arr[2, 31] = 0
-	arr[2, 32] = 0
-	arr[2, 33] = 0
-	arr[2, 34] = 0
-	arr[2, 35] = 0
-	arr[2, 36] = 0
-	arr[2, 37] = 0
-	arr[2, 38] = 0
-	arr[2, 39] = 0
-	arr[2, 40] = 0
-	arr[3, 0] = 0
-	arr[3, 1] = 0
-	arr[3, 2] = 0
-	arr[3, 3] = 0
-	arr[3, 4] = 170000.0*y[10]
-	arr[3, 5] = 0
-	arr[3, 6] = 0
-	arr[3, 7] = 0
-	arr[3, 8] = 0
-	arr[3, 9] = 0
-	arr[3, 10] = 170000.0*y[4]
-	arr[3, 11] = 0
-	arr[3, 12] = 0
-	arr[3, 13] = 0
-	arr[3, 14] = 0
-	arr[3, 15] = 0
-	arr[3, 16] = 0
-	arr[3, 17] = 0
-	arr[3, 18] = 0
-	arr[3, 19] = 0
-	arr[3, 20] = 0
-	arr[3, 21] = 0
-	arr[3, 22] = 0
-	arr[3, 23] = 0
-	arr[3, 24] = 0
-	arr[3, 25] = 0
-	arr[3, 26] = 0
-	arr[3, 27] = 0
-	arr[3, 28] = 0
-	arr[3, 29] = 0
-	arr[3, 30] = 0
-	arr[3, 31] = 0
-	arr[3, 32] = 0
-	arr[3, 33] = 0
-	arr[3, 34] = 0
-	arr[3, 35] = 0
-	arr[3, 36] = 0
-	arr[3, 37] = 0
-	arr[3, 38] = 0
-	arr[3, 39] = 0
-	arr[3, 40] = 0
-	arr[4, 0] = 0
-	arr[4, 1] = 0
-	arr[4, 2] = 0
-	arr[4, 3] = 0
-	arr[4, 4] = 0
-	arr[4, 5] = 0
-	arr[4, 6] = 0
-	arr[4, 7] = 0
-	arr[4, 8] = 0
-	arr[4, 9] = 100000000.0*y[12]
-	arr[4, 10] = 0
-	arr[4, 11] = 0
-	arr[4, 12] = 100000000.0*y[9]
-	arr[4, 13] = 0
-	arr[4, 14] = 0
-	arr[4, 15] = 0
-	arr[4, 16] = 0
-	arr[4, 17] = 0
-	arr[4, 18] = 0
-	arr[4, 19] = 0
-	arr[4, 20] = 0
-	arr[4, 21] = 0
-	arr[4, 22] = 0
-	arr[4, 23] = 0
-	arr[4, 24] = 0
-	arr[4, 25] = 0
-	arr[4, 26] = 0
-	arr[4, 27] = 0
-	arr[4, 28] = 0
-	arr[4, 29] = 0
-	arr[4, 30] = 0
-	arr[4, 31] = 0
-	arr[4, 32] = 0
-	arr[4, 33] = 0
-	arr[4, 34] = 0
-	arr[4, 35] = 0
-	arr[4, 36] = 0
-	arr[4, 37] = 0
-	arr[4, 38] = 0
-	arr[4, 39] = 0
-	arr[4, 40] = 0
-	arr[5, 0] = 0
-	arr[5, 1] = 0
-	arr[5, 2] = 0
-	arr[5, 3] = 0
-	arr[5, 4] = 0
-	arr[5, 5] = 0
-	arr[5, 6] = 0
-	arr[5, 7] = 0
-	arr[5, 8] = 0.01*y[11]
-	arr[5, 9] = 0
-	arr[5, 10] = 0
-	arr[5, 11] = 0.01*y[8]
-	arr[5, 12] = 0
-	arr[5, 13] = 0
-	arr[5, 14] = 0
-	arr[5, 15] = 0
-	arr[5, 16] = 0
-	arr[5, 17] = 0
-	arr[5, 18] = 0
-	arr[5, 19] = 0
-	arr[5, 20] = 0
-	arr[5, 21] = 0
-	arr[5, 22] = 0
-	arr[5, 23] = 0
-	arr[5, 24] = 0
-	arr[5, 25] = 0
-	arr[5, 26] = 0
-	arr[5, 27] = 0
-	arr[5, 28] = 0
-	arr[5, 29] = 0
-	arr[5, 30] = 0
-	arr[5, 31] = 0
-	arr[5, 32] = 0
-	arr[5, 33] = 0
-	arr[5, 34] = 0
-	arr[5, 35] = 0
-	arr[5, 36] = 0
-	arr[5, 37] = 0
-	arr[5, 38] = 0
-	arr[5, 39] = 0
-	arr[5, 40] = 0
-	arr[6, 0] = 0
-	arr[6, 1] = 0
-	arr[6, 2] = 0
-	arr[6, 3] = 0
-	arr[6, 4] = 0.0001*y[14]
-	arr[6, 5] = 0
-	arr[6, 6] = 0
-	arr[6, 7] = 0
-	arr[6, 8] = 0
-	arr[6, 9] = 0
-	arr[6, 10] = 0
-	arr[6, 11] = 0
-	arr[6, 12] = 0
-	arr[6, 13] = 0
-	arr[6, 14] = 0.0001*y[4]
-	arr[6, 15] = 0
-	arr[6, 16] = 0
-	arr[6, 17] = 0
-	arr[6, 18] = 0
-	arr[6, 19] = 0
-	arr[6, 20] = 0
-	arr[6, 21] = 0
-	arr[6, 22] = 0
-	arr[6, 23] = 0
-	arr[6, 24] = 0
-	arr[6, 25] = 0
-	arr[6, 26] = 0
-	arr[6, 27] = 0
-	arr[6, 28] = 0
-	arr[6, 29] = 0
-	arr[6, 30] = 0
-	arr[6, 31] = 0
-	arr[6, 32] = 0
-	arr[6, 33] = 0
-	arr[6, 34] = 0
-	arr[6, 35] = 0
-	arr[6, 36] = 0
-	arr[6, 37] = 0
-	arr[6, 38] = 0
-	arr[6, 39] = 0
-	arr[6, 40] = 0
-	arr[7, 0] = 0
-	arr[7, 1] = 0
-	arr[7, 2] = 0
-	arr[7, 3] = 0
-	arr[7, 4] = 0
-	arr[7, 5] = 0
-	arr[7, 6] = 0
-	arr[7, 7] = 0
-	arr[7, 8] = 0
-	arr[7, 9] = 0
-	arr[7, 10] = 0
-	arr[7, 11] = 0
-	arr[7, 12] = 100000000.0*y[13]
-	arr[7, 13] = 100000000.0*y[12]
-	arr[7, 14] = 0
-	arr[7, 15] = 0
-	arr[7, 16] = 0
-	arr[7, 17] = 0
-	arr[7, 18] = 0
-	arr[7, 19] = 0
-	arr[7, 20] = 0
-	arr[7, 21] = 0
-	arr[7, 22] = 0
-	arr[7, 23] = 0
-	arr[7, 24] = 0
-	arr[7, 25] = 0
-	arr[7, 26] = 0
-	arr[7, 27] = 0
-	arr[7, 28] = 0
-	arr[7, 29] = 0
-	arr[7, 30] = 0
-	arr[7, 31] = 0
-	arr[7, 32] = 0
-	arr[7, 33] = 0
-	arr[7, 34] = 0
-	arr[7, 35] = 0
-	arr[7, 36] = 0
-	arr[7, 37] = 0
-	arr[7, 38] = 0
-	arr[7, 39] = 0
-	arr[7, 40] = 0
-	arr[8, 0] = 0
-	arr[8, 1] = 0
-	arr[8, 2] = 0
-	arr[8, 3] = 0
-	arr[8, 4] = 170000.0*y[16]
-	arr[8, 5] = 0
-	arr[8, 6] = 0
-	arr[8, 7] = 0
-	arr[8, 8] = 0
-	arr[8, 9] = 0
-	arr[8, 10] = 0
-	arr[8, 11] = 0
-	arr[8, 12] = 0
-	arr[8, 13] = 0
-	arr[8, 14] = 0
-	arr[8, 15] = 0
-	arr[8, 16] = 170000.0*y[4]
-	arr[8, 17] = 0
-	arr[8, 18] = 0
-	arr[8, 19] = 0
-	arr[8, 20] = 0
-	arr[8, 21] = 0
-	arr[8, 22] = 0
-	arr[8, 23] = 0
-	arr[8, 24] = 0
-	arr[8, 25] = 0
-	arr[8, 26] = 0
-	arr[8, 27] = 0
-	arr[8, 28] = 0
-	arr[8, 29] = 0
-	arr[8, 30] = 0
-	arr[8, 31] = 0
-	arr[8, 32] = 0
-	arr[8, 33] = 0
-	arr[8, 34] = 0
-	arr[8, 35] = 0
-	arr[8, 36] = 0
-	arr[8, 37] = 0
-	arr[8, 38] = 0
-	arr[8, 39] = 0
-	arr[8, 40] = 0
-	arr[9, 0] = 0
-	arr[9, 1] = 0
-	arr[9, 2] = 0
-	arr[9, 3] = 0
-	arr[9, 4] = 0
-	arr[9, 5] = 0
-	arr[9, 6] = 0
-	arr[9, 7] = 0
-	arr[9, 8] = 0
-	arr[9, 9] = 0
-	arr[9, 10] = 0
-	arr[9, 11] = 0
-	arr[9, 12] = 0
-	arr[9, 13] = 0
-	arr[9, 14] = 0
-	arr[9, 15] = 100000000.0*y[18]
-	arr[9, 16] = 0
-	arr[9, 17] = 0
-	arr[9, 18] = 100000000.0*y[15]
-	arr[9, 19] = 0
-	arr[9, 20] = 0
-	arr[9, 21] = 0
-	arr[9, 22] = 0
-	arr[9, 23] = 0
-	arr[9, 24] = 0
-	arr[9, 25] = 0
-	arr[9, 26] = 0
-	arr[9, 27] = 0
-	arr[9, 28] = 0
-	arr[9, 29] = 0
-	arr[9, 30] = 0
-	arr[9, 31] = 0
-	arr[9, 32] = 0
-	arr[9, 33] = 0
-	arr[9, 34] = 0
-	arr[9, 35] = 0
-	arr[9, 36] = 0
-	arr[9, 37] = 0
-	arr[9, 38] = 0
-	arr[9, 39] = 0
-	arr[9, 40] = 0
-	arr[10, 0] = 0
-	arr[10, 1] = 0
-	arr[10, 2] = 0
-	arr[10, 3] = 0
-	arr[10, 4] = 0
-	arr[10, 5] = 0
-	arr[10, 6] = 0
-	arr[10, 7] = 0
-	arr[10, 8] = 0.01*y[17]
-	arr[10, 9] = 0
-	arr[10, 10] = 0
-	arr[10, 11] = 0
-	arr[10, 12] = 0
-	arr[10, 13] = 0
-	arr[10, 14] = 0
-	arr[10, 15] = 0
-	arr[10, 16] = 0
-	arr[10, 17] = 0.01*y[8]
-	arr[10, 18] = 0
-	arr[10, 19] = 0
-	arr[10, 20] = 0
-	arr[10, 21] = 0
-	arr[10, 22] = 0
-	arr[10, 23] = 0
-	arr[10, 24] = 0
-	arr[10, 25] = 0
-	arr[10, 26] = 0
-	arr[10, 27] = 0
-	arr[10, 28] = 0
-	arr[10, 29] = 0
-	arr[10, 30] = 0
-	arr[10, 31] = 0
-	arr[10, 32] = 0
-	arr[10, 33] = 0
-	arr[10, 34] = 0
-	arr[10, 35] = 0
-	arr[10, 36] = 0
-	arr[10, 37] = 0
-	arr[10, 38] = 0
-	arr[10, 39] = 0
-	arr[10, 40] = 0
-	arr[11, 0] = 0
-	arr[11, 1] = 0
-	arr[11, 2] = 0
-	arr[11, 3] = 0
-	arr[11, 4] = 0.0001*y[20]
-	arr[11, 5] = 0
-	arr[11, 6] = 0
-	arr[11, 7] = 0
-	arr[11, 8] = 0
-	arr[11, 9] = 0
-	arr[11, 10] = 0
-	arr[11, 11] = 0
-	arr[11, 12] = 0
-	arr[11, 13] = 0
-	arr[11, 14] = 0
-	arr[11, 15] = 0
-	arr[11, 16] = 0
-	arr[11, 17] = 0
-	arr[11, 18] = 0
-	arr[11, 19] = 0
-	arr[11, 20] = 0.0001*y[4]
-	arr[11, 21] = 0
-	arr[11, 22] = 0
-	arr[11, 23] = 0
-	arr[11, 24] = 0
-	arr[11, 25] = 0
-	arr[11, 26] = 0
-	arr[11, 27] = 0
-	arr[11, 28] = 0
-	arr[11, 29] = 0
-	arr[11, 30] = 0
-	arr[11, 31] = 0
-	arr[11, 32] = 0
-	arr[11, 33] = 0
-	arr[11, 34] = 0
-	arr[11, 35] = 0
-	arr[11, 36] = 0
-	arr[11, 37] = 0
-	arr[11, 38] = 0
-	arr[11, 39] = 0
-	arr[11, 40] = 0
-	arr[12, 0] = 0
-	arr[12, 1] = 0
-	arr[12, 2] = 0
-	arr[12, 3] = 0
-	arr[12, 4] = 0
-	arr[12, 5] = 0
-	arr[12, 6] = 0
-	arr[12, 7] = 0
-	arr[12, 8] = 0
-	arr[12, 9] = 0
-	arr[12, 10] = 0
-	arr[12, 11] = 0
-	arr[12, 12] = 0
-	arr[12, 13] = 0
-	arr[12, 14] = 0
-	arr[12, 15] = 0
-	arr[12, 16] = 0
-	arr[12, 17] = 0
-	arr[12, 18] = 100000000.0*y[19]
-	arr[12, 19] = 100000000.0*y[18]
-	arr[12, 20] = 0
-	arr[12, 21] = 0
-	arr[12, 22] = 0
-	arr[12, 23] = 0
-	arr[12, 24] = 0
-	arr[12, 25] = 0
-	arr[12, 26] = 0
-	arr[12, 27] = 0
-	arr[12, 28] = 0
-	arr[12, 29] = 0
-	arr[12, 30] = 0
-	arr[12, 31] = 0
-	arr[12, 32] = 0
-	arr[12, 33] = 0
-	arr[12, 34] = 0
-	arr[12, 35] = 0
-	arr[12, 36] = 0
-	arr[12, 37] = 0
-	arr[12, 38] = 0
-	arr[12, 39] = 0
-	arr[12, 40] = 0
-	arr[13, 0] = 0
-	arr[13, 1] = 0
-	arr[13, 2] = 0
-	arr[13, 3] = 0
-	arr[13, 4] = 170000.0*y[22]
-	arr[13, 5] = 0
-	arr[13, 6] = 0
-	arr[13, 7] = 0
-	arr[13, 8] = 0
-	arr[13, 9] = 0
-	arr[13, 10] = 0
-	arr[13, 11] = 0
-	arr[13, 12] = 0
-	arr[13, 13] = 0
-	arr[13, 14] = 0
-	arr[13, 15] = 0
-	arr[13, 16] = 0
-	arr[13, 17] = 0
-	arr[13, 18] = 0
-	arr[13, 19] = 0
-	arr[13, 20] = 0
-	arr[13, 21] = 0
-	arr[13, 22] = 170000.0*y[4]
-	arr[13, 23] = 0
-	arr[13, 24] = 0
-	arr[13, 25] = 0
-	arr[13, 26] = 0
-	arr[13, 27] = 0
-	arr[13, 28] = 0
-	arr[13, 29] = 0
-	arr[13, 30] = 0
-	arr[13, 31] = 0
-	arr[13, 32] = 0
-	arr[13, 33] = 0
-	arr[13, 34] = 0
-	arr[13, 35] = 0
-	arr[13, 36] = 0
-	arr[13, 37] = 0
-	arr[13, 38] = 0
-	arr[13, 39] = 0
-	arr[13, 40] = 0
-	arr[14, 0] = 0
-	arr[14, 1] = 0
-	arr[14, 2] = 0
-	arr[14, 3] = 0
-	arr[14, 4] = 0
-	arr[14, 5] = 0
-	arr[14, 6] = 0
-	arr[14, 7] = 0
-	arr[14, 8] = 0
-	arr[14, 9] = 0
-	arr[14, 10] = 0
-	arr[14, 11] = 0
-	arr[14, 12] = 0
-	arr[14, 13] = 0
-	arr[14, 14] = 0
-	arr[14, 15] = 0
-	arr[14, 16] = 0
-	arr[14, 17] = 0
-	arr[14, 18] = 0
-	arr[14, 19] = 0
-	arr[14, 20] = 0
-	arr[14, 21] = 100000000.0*y[24]
-	arr[14, 22] = 0
-	arr[14, 23] = 0
-	arr[14, 24] = 100000000.0*y[21]
-	arr[14, 25] = 0
-	arr[14, 26] = 0
-	arr[14, 27] = 0
-	arr[14, 28] = 0
-	arr[14, 29] = 0
-	arr[14, 30] = 0
-	arr[14, 31] = 0
-	arr[14, 32] = 0
-	arr[14, 33] = 0
-	arr[14, 34] = 0
-	arr[14, 35] = 0
-	arr[14, 36] = 0
-	arr[14, 37] = 0
-	arr[14, 38] = 0
-	arr[14, 39] = 0
-	arr[14, 40] = 0
-	arr[15, 0] = 0
-	arr[15, 1] = 0
-	arr[15, 2] = 0
-	arr[15, 3] = 0
-	arr[15, 4] = 0
-	arr[15, 5] = 0
-	arr[15, 6] = 0
-	arr[15, 7] = 0
-	arr[15, 8] = 0.01*y[23]
-	arr[15, 9] = 0
-	arr[15, 10] = 0
-	arr[15, 11] = 0
-	arr[15, 12] = 0
-	arr[15, 13] = 0
-	arr[15, 14] = 0
-	arr[15, 15] = 0
-	arr[15, 16] = 0
-	arr[15, 17] = 0
-	arr[15, 18] = 0
-	arr[15, 19] = 0
-	arr[15, 20] = 0
-	arr[15, 21] = 0
-	arr[15, 22] = 0
-	arr[15, 23] = 0.01*y[8]
-	arr[15, 24] = 0
-	arr[15, 25] = 0
-	arr[15, 26] = 0
-	arr[15, 27] = 0
-	arr[15, 28] = 0
-	arr[15, 29] = 0
-	arr[15, 30] = 0
-	arr[15, 31] = 0
-	arr[15, 32] = 0
-	arr[15, 33] = 0
-	arr[15, 34] = 0
-	arr[15, 35] = 0
-	arr[15, 36] = 0
-	arr[15, 37] = 0
-	arr[15, 38] = 0
-	arr[15, 39] = 0
-	arr[15, 40] = 0
-	arr[16, 0] = 0
-	arr[16, 1] = 0
-	arr[16, 2] = 0
-	arr[16, 3] = 0
-	arr[16, 4] = 0.0001*y[26]
-	arr[16, 5] = 0
-	arr[16, 6] = 0
-	arr[16, 7] = 0
-	arr[16, 8] = 0
-	arr[16, 9] = 0
-	arr[16, 10] = 0
-	arr[16, 11] = 0
-	arr[16, 12] = 0
-	arr[16, 13] = 0
-	arr[16, 14] = 0
-	arr[16, 15] = 0
-	arr[16, 16] = 0
-	arr[16, 17] = 0
-	arr[16, 18] = 0
-	arr[16, 19] = 0
-	arr[16, 20] = 0
-	arr[16, 21] = 0
-	arr[16, 22] = 0
-	arr[16, 23] = 0
-	arr[16, 24] = 0
-	arr[16, 25] = 0
-	arr[16, 26] = 0.0001*y[4]
-	arr[16, 27] = 0
-	arr[16, 28] = 0
-	arr[16, 29] = 0
-	arr[16, 30] = 0
-	arr[16, 31] = 0
-	arr[16, 32] = 0
-	arr[16, 33] = 0
-	arr[16, 34] = 0
-	arr[16, 35] = 0
-	arr[16, 36] = 0
-	arr[16, 37] = 0
-	arr[16, 38] = 0
-	arr[16, 39] = 0
-	arr[16, 40] = 0
-	arr[17, 0] = 0
-	arr[17, 1] = 0
-	arr[17, 2] = 0
-	arr[17, 3] = 0
-	arr[17, 4] = 0
-	arr[17, 5] = 0
-	arr[17, 6] = 0
-	arr[17, 7] = 0
-	arr[17, 8] = 0
-	arr[17, 9] = 0
-	arr[17, 10] = 0
-	arr[17, 11] = 0
-	arr[17, 12] = 0
-	arr[17, 13] = 0
-	arr[17, 14] = 0
-	arr[17, 15] = 0
-	arr[17, 16] = 0
-	arr[17, 17] = 0
-	arr[17, 18] = 0
-	arr[17, 19] = 0
-	arr[17, 20] = 0
-	arr[17, 21] = 0
-	arr[17, 22] = 0
-	arr[17, 23] = 0
-	arr[17, 24] = 100000000.0*y[25]
-	arr[17, 25] = 100000000.0*y[24]
-	arr[17, 26] = 0
-	arr[17, 27] = 0
-	arr[17, 28] = 0
-	arr[17, 29] = 0
-	arr[17, 30] = 0
-	arr[17, 31] = 0
-	arr[17, 32] = 0
-	arr[17, 33] = 0
-	arr[17, 34] = 0
-	arr[17, 35] = 0
-	arr[17, 36] = 0
-	arr[17, 37] = 0
-	arr[17, 38] = 0
-	arr[17, 39] = 0
-	arr[17, 40] = 0
-	arr[18, 0] = 0
-	arr[18, 1] = 0
-	arr[18, 2] = 0
-	arr[18, 3] = 0
-	arr[18, 4] = 170000.0*y[28]
-	arr[18, 5] = 0
-	arr[18, 6] = 0
-	arr[18, 7] = 0
-	arr[18, 8] = 0
-	arr[18, 9] = 0
-	arr[18, 10] = 0
-	arr[18, 11] = 0
-	arr[18, 12] = 0
-	arr[18, 13] = 0
-	arr[18, 14] = 0
-	arr[18, 15] = 0
-	arr[18, 16] = 0
-	arr[18, 17] = 0
-	arr[18, 18] = 0
-	arr[18, 19] = 0
-	arr[18, 20] = 0
-	arr[18, 21] = 0
-	arr[18, 22] = 0
-	arr[18, 23] = 0
-	arr[18, 24] = 0
-	arr[18, 25] = 0
-	arr[18, 26] = 0
-	arr[18, 27] = 0
-	arr[18, 28] = 170000.0*y[4]
-	arr[18, 29] = 0
-	arr[18, 30] = 0
-	arr[18, 31] = 0
-	arr[18, 32] = 0
-	arr[18, 33] = 0
-	arr[18, 34] = 0
-	arr[18, 35] = 0
-	arr[18, 36] = 0
-	arr[18, 37] = 0
-	arr[18, 38] = 0
-	arr[18, 39] = 0
-	arr[18, 40] = 0
-	arr[19, 0] = 0
-	arr[19, 1] = 0
-	arr[19, 2] = 0
-	arr[19, 3] = 0
-	arr[19, 4] = 0
-	arr[19, 5] = 0
-	arr[19, 6] = 0
-	arr[19, 7] = 0
-	arr[19, 8] = 0
-	arr[19, 9] = 0
-	arr[19, 10] = 0
-	arr[19, 11] = 0
-	arr[19, 12] = 0
-	arr[19, 13] = 0
-	arr[19, 14] = 0
-	arr[19, 15] = 0
-	arr[19, 16] = 0
-	arr[19, 17] = 0
-	arr[19, 18] = 0
-	arr[19, 19] = 0
-	arr[19, 20] = 0
-	arr[19, 21] = 0
-	arr[19, 22] = 0
-	arr[19, 23] = 0
-	arr[19, 24] = 0
-	arr[19, 25] = 0
-	arr[19, 26] = 0
-	arr[19, 27] = 100000000.0*y[30]
-	arr[19, 28] = 0
-	arr[19, 29] = 0
-	arr[19, 30] = 100000000.0*y[27]
-	arr[19, 31] = 0
-	arr[19, 32] = 0
-	arr[19, 33] = 0
-	arr[19, 34] = 0
-	arr[19, 35] = 0
-	arr[19, 36] = 0
-	arr[19, 37] = 0
-	arr[19, 38] = 0
-	arr[19, 39] = 0
-	arr[19, 40] = 0
-	arr[20, 0] = 0
-	arr[20, 1] = 0
-	arr[20, 2] = 0
-	arr[20, 3] = 0
-	arr[20, 4] = 0
-	arr[20, 5] = 0
-	arr[20, 6] = 0
-	arr[20, 7] = 0
-	arr[20, 8] = 0.01*y[29]
-	arr[20, 9] = 0
-	arr[20, 10] = 0
-	arr[20, 11] = 0
-	arr[20, 12] = 0
-	arr[20, 13] = 0
-	arr[20, 14] = 0
-	arr[20, 15] = 0
-	arr[20, 16] = 0
-	arr[20, 17] = 0
-	arr[20, 18] = 0
-	arr[20, 19] = 0
-	arr[20, 20] = 0
-	arr[20, 21] = 0
-	arr[20, 22] = 0
-	arr[20, 23] = 0
-	arr[20, 24] = 0
-	arr[20, 25] = 0
-	arr[20, 26] = 0
-	arr[20, 27] = 0
-	arr[20, 28] = 0
-	arr[20, 29] = 0.01*y[8]
-	arr[20, 30] = 0
-	arr[20, 31] = 0
-	arr[20, 32] = 0
-	arr[20, 33] = 0
-	arr[20, 34] = 0
-	arr[20, 35] = 0
-	arr[20, 36] = 0
-	arr[20, 37] = 0
-	arr[20, 38] = 0
-	arr[20, 39] = 0
-	arr[20, 40] = 0
-	arr[21, 0] = 0
-	arr[21, 1] = 0
-	arr[21, 2] = 0
-	arr[21, 3] = 0
-	arr[21, 4] = 0.0001*y[32]
-	arr[21, 5] = 0
-	arr[21, 6] = 0
-	arr[21, 7] = 0
-	arr[21, 8] = 0
-	arr[21, 9] = 0
-	arr[21, 10] = 0
-	arr[21, 11] = 0
-	arr[21, 12] = 0
-	arr[21, 13] = 0
-	arr[21, 14] = 0
-	arr[21, 15] = 0
-	arr[21, 16] = 0
-	arr[21, 17] = 0
-	arr[21, 18] = 0
-	arr[21, 19] = 0
-	arr[21, 20] = 0
-	arr[21, 21] = 0
-	arr[21, 22] = 0
-	arr[21, 23] = 0
-	arr[21, 24] = 0
-	arr[21, 25] = 0
-	arr[21, 26] = 0
-	arr[21, 27] = 0
-	arr[21, 28] = 0
-	arr[21, 29] = 0
-	arr[21, 30] = 0
-	arr[21, 31] = 0
-	arr[21, 32] = 0.0001*y[4]
-	arr[21, 33] = 0
-	arr[21, 34] = 0
-	arr[21, 35] = 0
-	arr[21, 36] = 0
-	arr[21, 37] = 0
-	arr[21, 38] = 0
-	arr[21, 39] = 0
-	arr[21, 40] = 0
-	arr[22, 0] = 0
-	arr[22, 1] = 0
-	arr[22, 2] = 0
-	arr[22, 3] = 0
-	arr[22, 4] = 0
-	arr[22, 5] = 0
-	arr[22, 6] = 0
-	arr[22, 7] = 0
-	arr[22, 8] = 0
-	arr[22, 9] = 0
-	arr[22, 10] = 0
-	arr[22, 11] = 0
-	arr[22, 12] = 0
-	arr[22, 13] = 0
-	arr[22, 14] = 0
-	arr[22, 15] = 0
-	arr[22, 16] = 0
-	arr[22, 17] = 0
-	arr[22, 18] = 0
-	arr[22, 19] = 0
-	arr[22, 20] = 0
-	arr[22, 21] = 0
-	arr[22, 22] = 0
-	arr[22, 23] = 0
-	arr[22, 24] = 0
-	arr[22, 25] = 0
-	arr[22, 26] = 0
-	arr[22, 27] = 0
-	arr[22, 28] = 0
-	arr[22, 29] = 0
-	arr[22, 30] = 100000000.0*y[31]
-	arr[22, 31] = 100000000.0*y[30]
-	arr[22, 32] = 0
-	arr[22, 33] = 0
-	arr[22, 34] = 0
-	arr[22, 35] = 0
-	arr[22, 36] = 0
-	arr[22, 37] = 0
-	arr[22, 38] = 0
-	arr[22, 39] = 0
-	arr[22, 40] = 0
-	arr[23, 0] = 0
-	arr[23, 1] = 0
-	arr[23, 2] = 0
-	arr[23, 3] = 0
-	arr[23, 4] = 500.0*y[34]
-	arr[23, 5] = 0
-	arr[23, 6] = 0
-	arr[23, 7] = 0
-	arr[23, 8] = 0
-	arr[23, 9] = 0
-	arr[23, 10] = 0
-	arr[23, 11] = 0
-	arr[23, 12] = 0
-	arr[23, 13] = 0
-	arr[23, 14] = 0
-	arr[23, 15] = 0
-	arr[23, 16] = 0
-	arr[23, 17] = 0
-	arr[23, 18] = 0
-	arr[23, 19] = 0
-	arr[23, 20] = 0
-	arr[23, 21] = 0
-	arr[23, 22] = 0
-	arr[23, 23] = 0
-	arr[23, 24] = 0
-	arr[23, 25] = 0
-	arr[23, 26] = 0
-	arr[23, 27] = 0
-	arr[23, 28] = 0
-	arr[23, 29] = 0
-	arr[23, 30] = 0
-	arr[23, 31] = 0
-	arr[23, 32] = 0
-	arr[23, 33] = 0
-	arr[23, 34] = 500.0*y[4]
-	arr[23, 35] = 0
-	arr[23, 36] = 0
-	arr[23, 37] = 0
-	arr[23, 38] = 0
-	arr[23, 39] = 0
-	arr[23, 40] = 0
-	arr[24, 0] = 0
-	arr[24, 1] = 0
-	arr[24, 2] = 0
-	arr[24, 3] = 0
-	arr[24, 4] = 0
-	arr[24, 5] = 0
-	arr[24, 6] = 0
-	arr[24, 7] = 0
-	arr[24, 8] = 0
-	arr[24, 9] = 0
-	arr[24, 10] = 0
-	arr[24, 11] = 0
-	arr[24, 12] = 0
-	arr[24, 13] = 0
-	arr[24, 14] = 0
-	arr[24, 15] = 0
-	arr[24, 16] = 0
-	arr[24, 17] = 0
-	arr[24, 18] = 0
-	arr[24, 19] = 0
-	arr[24, 20] = 0
-	arr[24, 21] = 0
-	arr[24, 22] = 0
-	arr[24, 23] = 0
-	arr[24, 24] = 0
-	arr[24, 25] = 0
-	arr[24, 26] = 0
-	arr[24, 27] = 0
-	arr[24, 28] = 0
-	arr[24, 29] = 0
-	arr[24, 30] = 0
-	arr[24, 31] = 0
-	arr[24, 32] = 0
-	arr[24, 33] = 100000000.0*y[36]
-	arr[24, 34] = 0
-	arr[24, 35] = 0
-	arr[24, 36] = 100000000.0*y[33]
-	arr[24, 37] = 0
-	arr[24, 38] = 0
-	arr[24, 39] = 0
-	arr[24, 40] = 0
-	arr[25, 0] = 0
-	arr[25, 1] = 0
-	arr[25, 2] = 0
-	arr[25, 3] = 0
-	arr[25, 4] = 0
-	arr[25, 5] = 0
-	arr[25, 6] = 0
-	arr[25, 7] = 0
-	arr[25, 8] = 0.01*y[35]
-	arr[25, 9] = 0
-	arr[25, 10] = 0
-	arr[25, 11] = 0
-	arr[25, 12] = 0
-	arr[25, 13] = 0
-	arr[25, 14] = 0
-	arr[25, 15] = 0
-	arr[25, 16] = 0
-	arr[25, 17] = 0
-	arr[25, 18] = 0
-	arr[25, 19] = 0
-	arr[25, 20] = 0
-	arr[25, 21] = 0
-	arr[25, 22] = 0
-	arr[25, 23] = 0
-	arr[25, 24] = 0
-	arr[25, 25] = 0
-	arr[25, 26] = 0
-	arr[25, 27] = 0
-	arr[25, 28] = 0
-	arr[25, 29] = 0
-	arr[25, 30] = 0
-	arr[25, 31] = 0
-	arr[25, 32] = 0
-	arr[25, 33] = 0
-	arr[25, 34] = 0
-	arr[25, 35] = 0.01*y[8]
-	arr[25, 36] = 0
-	arr[25, 37] = 0
-	arr[25, 38] = 0
-	arr[25, 39] = 0
-	arr[25, 40] = 0
-	arr[26, 0] = 0
-	arr[26, 1] = 0
-	arr[26, 2] = 0
-	arr[26, 3] = 0
-	arr[26, 4] = 500.0*y[38]
-	arr[26, 5] = 0
-	arr[26, 6] = 0
-	arr[26, 7] = 0
-	arr[26, 8] = 0
-	arr[26, 9] = 0
-	arr[26, 10] = 0
-	arr[26, 11] = 0
-	arr[26, 12] = 0
-	arr[26, 13] = 0
-	arr[26, 14] = 0
-	arr[26, 15] = 0
-	arr[26, 16] = 0
-	arr[26, 17] = 0
-	arr[26, 18] = 0
-	arr[26, 19] = 0
-	arr[26, 20] = 0
-	arr[26, 21] = 0
-	arr[26, 22] = 0
-	arr[26, 23] = 0
-	arr[26, 24] = 0
-	arr[26, 25] = 0
-	arr[26, 26] = 0
-	arr[26, 27] = 0
-	arr[26, 28] = 0
-	arr[26, 29] = 0
-	arr[26, 30] = 0
-	arr[26, 31] = 0
-	arr[26, 32] = 0
-	arr[26, 33] = 0
-	arr[26, 34] = 0
-	arr[26, 35] = 0
-	arr[26, 36] = 0
-	arr[26, 37] = 0
-	arr[26, 38] = 500.0*y[4]
-	arr[26, 39] = 0
-	arr[26, 40] = 0
-	arr[27, 0] = 0
-	arr[27, 1] = 0
-	arr[27, 2] = 0
-	arr[27, 3] = 0
-	arr[27, 4] = 0
-	arr[27, 5] = 0
-	arr[27, 6] = 0
-	arr[27, 7] = 0
-	arr[27, 8] = 0
-	arr[27, 9] = 0
-	arr[27, 10] = 0
-	arr[27, 11] = 0
-	arr[27, 12] = 0
-	arr[27, 13] = 0
-	arr[27, 14] = 0
-	arr[27, 15] = 0
-	arr[27, 16] = 0
-	arr[27, 17] = 0
-	arr[27, 18] = 0
-	arr[27, 19] = 0
-	arr[27, 20] = 0
-	arr[27, 21] = 0
-	arr[27, 22] = 0
-	arr[27, 23] = 0
-	arr[27, 24] = 0
-	arr[27, 25] = 0
-	arr[27, 26] = 0
-	arr[27, 27] = 0
-	arr[27, 28] = 0
-	arr[27, 29] = 0
-	arr[27, 30] = 0
-	arr[27, 31] = 0
-	arr[27, 32] = 0
-	arr[27, 33] = 0
-	arr[27, 34] = 0
-	arr[27, 35] = 0
-	arr[27, 36] = 0
-	arr[27, 37] = 100000000.0*y[40]
-	arr[27, 38] = 0
-	arr[27, 39] = 0
-	arr[27, 40] = 100000000.0*y[37]
-	arr[28, 0] = 0
-	arr[28, 1] = 0
-	arr[28, 2] = 0
-	arr[28, 3] = 0
-	arr[28, 4] = 0
-	arr[28, 5] = 0
-	arr[28, 6] = 0
-	arr[28, 7] = 0
-	arr[28, 8] = 0.01*y[39]
-	arr[28, 9] = 0
-	arr[28, 10] = 0
-	arr[28, 11] = 0
-	arr[28, 12] = 0
-	arr[28, 13] = 0
-	arr[28, 14] = 0
-	arr[28, 15] = 0
-	arr[28, 16] = 0
-	arr[28, 17] = 0
-	arr[28, 18] = 0
-	arr[28, 19] = 0
-	arr[28, 20] = 0
-	arr[28, 21] = 0
-	arr[28, 22] = 0
-	arr[28, 23] = 0
-	arr[28, 24] = 0
-	arr[28, 25] = 0
-	arr[28, 26] = 0
-	arr[28, 27] = 0
-	arr[28, 28] = 0
-	arr[28, 29] = 0
-	arr[28, 30] = 0
-	arr[28, 31] = 0
-	arr[28, 32] = 0
-	arr[28, 33] = 0
-	arr[28, 34] = 0
-	arr[28, 35] = 0
-	arr[28, 36] = 0
-	arr[28, 37] = 0
-	arr[28, 38] = 0
-	arr[28, 39] = 0.01*y[8]
-	arr[28, 40] = 0
-	return arr.reshape(-1)
-
-def eq_rates(t, y, kf, kr):
-	arr = np.zeros((37, 1))
-	arr[0, 0] = kf[0]*y[1]*y[2] - kr[0]*y[0]
-	arr[1, 0] = kf[1]*y[4]*y[5] - kr[1]*y[3]
-	arr[2, 0] = kf[2]*y[7]*y[8] - kr[2]*y[6]
-	arr[3, 0] = kf[3]*y[10]*y[11] - kr[3]*y[9]
-	arr[4, 0] = kf[4]*y[10]*y[13] - kr[4]*y[12]
-	arr[5, 0] = kf[5]*y[15]*y[16] - kr[5]*y[14]
-	arr[6, 0] = kf[6]*y[18]*y[19] - kr[6]*y[17]
-	arr[7, 0] = kf[7]*y[21]*y[22] - kr[7]*y[20]
-	arr[8, 0] = kf[8]*y[24]*y[25] - kr[8]*y[23]
-	arr[9, 0] = kf[9]*y[27]*y[28] - kr[9]*y[26]
-	arr[10, 0] = kf[10]*y[30]*y[31] - kr[10]*y[29]
-	arr[11, 0] = kf[11]*y[33]*y[34] - kr[11]*y[32]
-	arr[12, 0] = kf[12]*y[36]*y[37] - kr[12]*y[35]
-	arr[13, 0] = kf[13]*y[39]*y[40] - kr[13]*y[38]
-	arr[14, 0] = kf[14]*y[22]*y[42] - kr[14]*y[41]
-	arr[15, 0] = kf[15]*y[10]*y[44] - kr[15]*y[43]
-	arr[16, 0] = kf[16]*y[46]*y[47] - kr[16]*y[45]
-	arr[17, 0] = kf[17]*y[49]*y[50] - kr[17]*y[48]
-	arr[18, 0] = kf[18]*y[52]*y[53] - kr[18]*y[51]
-	arr[19, 0] = kf[19]*y[55]*y[56] - kr[19]*y[54]
-	arr[20, 0] = kf[20]*y[22]*y[58] - kr[20]*y[57]
-	arr[21, 0] = kf[21]*y[60]*y[61] - kr[21]*y[59]
-	arr[22, 0] = kf[22]*y[63]*y[64] - kr[22]*y[62]
-	arr[23, 0] = kf[23]*y[66]*y[67] - kr[23]*y[65]
-	arr[24, 0] = kf[24]*y[66]*y[69] - kr[24]*y[68]
-	arr[25, 0] = kf[25]*y[63]*y[71] - kr[25]*y[70]
-	arr[26, 0] = kf[26]*y[40]*y[73] - kr[26]*y[72]
-	arr[27, 0] = kf[27]*y[75]*y[76]**2.0 - kr[27]**2.0*y[74]
-	arr[28, 0] = kf[28]*y[78]*y[79] - kr[28]*y[77]
-	arr[29, 0] = kf[29]*y[81]*y[82]**2.0 - kr[29]**2.0*y[80]
-	arr[30, 0] = kf[30]*y[84]*y[85] - kr[30]*y[83]
-	arr[31, 0] = kf[31]*y[87]*y[88] - kr[31]*y[86]
-	arr[32, 0] = kf[32]*y[90]*y[91] - kr[32]*y[89]
-	arr[33, 0] = kf[33]*y[93]*y[94] - kr[33]*y[92]
-	arr[34, 0] = kf[34]*y[96]*y[97] - kr[34]*y[95]
-	arr[35, 0] = kf[35]*y[100]*y[99] - kr[35]*y[98]
-	arr[36, 0] = kf[36]*y[102]*y[103] - kr[36]*y[101]
-	return arr.reshape(-1)
-
-def eq_rates_jac(t, y, kf, kr):
-	arr = np.zeros((37, 104))
-	arr[0, 0] = -kr[0]
-	arr[0, 1] = kf[0] * y[2]
-	arr[0, 2] = kf[0] * y[1]
-	arr[0, 3] = 0
-	arr[0, 4] = 0
-	arr[0, 5] = 0
-	arr[0, 6] = 0
-	arr[0, 7] = 0
-	arr[0, 8] = 0
-	arr[0, 9] = 0
-	arr[0, 10] = 0
-	arr[0, 11] = 0
-	arr[0, 12] = 0
-	arr[0, 13] = 0
-	arr[0, 14] = 0
-	arr[0, 15] = 0
-	arr[0, 16] = 0
-	arr[0, 17] = 0
-	arr[0, 18] = 0
-	arr[0, 19] = 0
-	arr[0, 20] = 0
-	arr[0, 21] = 0
-	arr[0, 22] = 0
-	arr[0, 23] = 0
-	arr[0, 24] = 0
-	arr[0, 25] = 0
-	arr[0, 26] = 0
-	arr[0, 27] = 0
-	arr[0, 28] = 0
-	arr[0, 29] = 0
-	arr[0, 30] = 0
-	arr[0, 31] = 0
-	arr[0, 32] = 0
-	arr[0, 33] = 0
-	arr[0, 34] = 0
-	arr[0, 35] = 0
-	arr[0, 36] = 0
-	arr[0, 37] = 0
-	arr[0, 38] = 0
-	arr[0, 39] = 0
-	arr[0, 40] = 0
-	arr[0, 41] = 0
-	arr[0, 42] = 0
-	arr[0, 43] = 0
-	arr[0, 44] = 0
-	arr[0, 45] = 0
-	arr[0, 46] = 0
-	arr[0, 47] = 0
-	arr[0, 48] = 0
-	arr[0, 49] = 0
-	arr[0, 50] = 0
-	arr[0, 51] = 0
-	arr[0, 52] = 0
-	arr[0, 53] = 0
-	arr[0, 54] = 0
-	arr[0, 55] = 0
-	arr[0, 56] = 0
-	arr[0, 57] = 0
-	arr[0, 58] = 0
-	arr[0, 59] = 0
-	arr[0, 60] = 0
-	arr[0, 61] = 0
-	arr[0, 62] = 0
-	arr[0, 63] = 0
-	arr[0, 64] = 0
-	arr[0, 65] = 0
-	arr[0, 66] = 0
-	arr[0, 67] = 0
-	arr[0, 68] = 0
-	arr[0, 69] = 0
-	arr[0, 70] = 0
-	arr[0, 71] = 0
-	arr[0, 72] = 0
-	arr[0, 73] = 0
-	arr[0, 74] = 0
-	arr[0, 75] = 0
-	arr[0, 76] = 0
-	arr[0, 77] = 0
-	arr[0, 78] = 0
-	arr[0, 79] = 0
-	arr[0, 80] = 0
-	arr[0, 81] = 0
-	arr[0, 82] = 0
-	arr[0, 83] = 0
-	arr[0, 84] = 0
-	arr[0, 85] = 0
-	arr[0, 86] = 0
-	arr[0, 87] = 0
-	arr[0, 88] = 0
-	arr[0, 89] = 0
-	arr[0, 90] = 0
-	arr[0, 91] = 0
-	arr[0, 92] = 0
-	arr[0, 93] = 0
-	arr[0, 94] = 0
-	arr[0, 95] = 0
-	arr[0, 96] = 0
-	arr[0, 97] = 0
-	arr[0, 98] = 0
-	arr[0, 99] = 0
-	arr[0, 100] = 0
-	arr[0, 101] = 0
-	arr[0, 102] = 0
-	arr[0, 103] = 0
-	arr[1, 0] = 0
-	arr[1, 1] = 0
-	arr[1, 2] = 0
-	arr[1, 3] = -kr[1]
-	arr[1, 4] = kf[1] * y[5]
-	arr[1, 5] = kf[1] * y[4]
-	arr[1, 6] = 0
-	arr[1, 7] = 0
-	arr[1, 8] = 0
-	arr[1, 9] = 0
-	arr[1, 10] = 0
-	arr[1, 11] = 0
-	arr[1, 12] = 0
-	arr[1, 13] = 0
-	arr[1, 14] = 0
-	arr[1, 15] = 0
-	arr[1, 16] = 0
-	arr[1, 17] = 0
-	arr[1, 18] = 0
-	arr[1, 19] = 0
-	arr[1, 20] = 0
-	arr[1, 21] = 0
-	arr[1, 22] = 0
-	arr[1, 23] = 0
-	arr[1, 24] = 0
-	arr[1, 25] = 0
-	arr[1, 26] = 0
-	arr[1, 27] = 0
-	arr[1, 28] = 0
-	arr[1, 29] = 0
-	arr[1, 30] = 0
-	arr[1, 31] = 0
-	arr[1, 32] = 0
-	arr[1, 33] = 0
-	arr[1, 34] = 0
-	arr[1, 35] = 0
-	arr[1, 36] = 0
-	arr[1, 37] = 0
-	arr[1, 38] = 0
-	arr[1, 39] = 0
-	arr[1, 40] = 0
-	arr[1, 41] = 0
-	arr[1, 42] = 0
-	arr[1, 43] = 0
-	arr[1, 44] = 0
-	arr[1, 45] = 0
-	arr[1, 46] = 0
-	arr[1, 47] = 0
-	arr[1, 48] = 0
-	arr[1, 49] = 0
-	arr[1, 50] = 0
-	arr[1, 51] = 0
-	arr[1, 52] = 0
-	arr[1, 53] = 0
-	arr[1, 54] = 0
-	arr[1, 55] = 0
-	arr[1, 56] = 0
-	arr[1, 57] = 0
-	arr[1, 58] = 0
-	arr[1, 59] = 0
-	arr[1, 60] = 0
-	arr[1, 61] = 0
-	arr[1, 62] = 0
-	arr[1, 63] = 0
-	arr[1, 64] = 0
-	arr[1, 65] = 0
-	arr[1, 66] = 0
-	arr[1, 67] = 0
-	arr[1, 68] = 0
-	arr[1, 69] = 0
-	arr[1, 70] = 0
-	arr[1, 71] = 0
-	arr[1, 72] = 0
-	arr[1, 73] = 0
-	arr[1, 74] = 0
-	arr[1, 75] = 0
-	arr[1, 76] = 0
-	arr[1, 77] = 0
-	arr[1, 78] = 0
-	arr[1, 79] = 0
-	arr[1, 80] = 0
-	arr[1, 81] = 0
-	arr[1, 82] = 0
-	arr[1, 83] = 0
-	arr[1, 84] = 0
-	arr[1, 85] = 0
-	arr[1, 86] = 0
-	arr[1, 87] = 0
-	arr[1, 88] = 0
-	arr[1, 89] = 0
-	arr[1, 90] = 0
-	arr[1, 91] = 0
-	arr[1, 92] = 0
-	arr[1, 93] = 0
-	arr[1, 94] = 0
-	arr[1, 95] = 0
-	arr[1, 96] = 0
-	arr[1, 97] = 0
-	arr[1, 98] = 0
-	arr[1, 99] = 0
-	arr[1, 100] = 0
-	arr[1, 101] = 0
-	arr[1, 102] = 0
-	arr[1, 103] = 0
-	arr[2, 0] = 0
-	arr[2, 1] = 0
-	arr[2, 2] = 0
-	arr[2, 3] = 0
-	arr[2, 4] = 0
-	arr[2, 5] = 0
-	arr[2, 6] = -kr[2]
-	arr[2, 7] = kf[2] * y[8]
-	arr[2, 8] = kf[2] * y[7]
-	arr[2, 9] = 0
-	arr[2, 10] = 0
-	arr[2, 11] = 0
-	arr[2, 12] = 0
-	arr[2, 13] = 0
-	arr[2, 14] = 0
-	arr[2, 15] = 0
-	arr[2, 16] = 0
-	arr[2, 17] = 0
-	arr[2, 18] = 0
-	arr[2, 19] = 0
-	arr[2, 20] = 0
-	arr[2, 21] = 0
-	arr[2, 22] = 0
-	arr[2, 23] = 0
-	arr[2, 24] = 0
-	arr[2, 25] = 0
-	arr[2, 26] = 0
-	arr[2, 27] = 0
-	arr[2, 28] = 0
-	arr[2, 29] = 0
-	arr[2, 30] = 0
-	arr[2, 31] = 0
-	arr[2, 32] = 0
-	arr[2, 33] = 0
-	arr[2, 34] = 0
-	arr[2, 35] = 0
-	arr[2, 36] = 0
-	arr[2, 37] = 0
-	arr[2, 38] = 0
-	arr[2, 39] = 0
-	arr[2, 40] = 0
-	arr[2, 41] = 0
-	arr[2, 42] = 0
-	arr[2, 43] = 0
-	arr[2, 44] = 0
-	arr[2, 45] = 0
-	arr[2, 46] = 0
-	arr[2, 47] = 0
-	arr[2, 48] = 0
-	arr[2, 49] = 0
-	arr[2, 50] = 0
-	arr[2, 51] = 0
-	arr[2, 52] = 0
-	arr[2, 53] = 0
-	arr[2, 54] = 0
-	arr[2, 55] = 0
-	arr[2, 56] = 0
-	arr[2, 57] = 0
-	arr[2, 58] = 0
-	arr[2, 59] = 0
-	arr[2, 60] = 0
-	arr[2, 61] = 0
-	arr[2, 62] = 0
-	arr[2, 63] = 0
-	arr[2, 64] = 0
-	arr[2, 65] = 0
-	arr[2, 66] = 0
-	arr[2, 67] = 0
-	arr[2, 68] = 0
-	arr[2, 69] = 0
-	arr[2, 70] = 0
-	arr[2, 71] = 0
-	arr[2, 72] = 0
-	arr[2, 73] = 0
-	arr[2, 74] = 0
-	arr[2, 75] = 0
-	arr[2, 76] = 0
-	arr[2, 77] = 0
-	arr[2, 78] = 0
-	arr[2, 79] = 0
-	arr[2, 80] = 0
-	arr[2, 81] = 0
-	arr[2, 82] = 0
-	arr[2, 83] = 0
-	arr[2, 84] = 0
-	arr[2, 85] = 0
-	arr[2, 86] = 0
-	arr[2, 87] = 0
-	arr[2, 88] = 0
-	arr[2, 89] = 0
-	arr[2, 90] = 0
-	arr[2, 91] = 0
-	arr[2, 92] = 0
-	arr[2, 93] = 0
-	arr[2, 94] = 0
-	arr[2, 95] = 0
-	arr[2, 96] = 0
-	arr[2, 97] = 0
-	arr[2, 98] = 0
-	arr[2, 99] = 0
-	arr[2, 100] = 0
-	arr[2, 101] = 0
-	arr[2, 102] = 0
-	arr[2, 103] = 0
-	arr[3, 0] = 0
-	arr[3, 1] = 0
-	arr[3, 2] = 0
-	arr[3, 3] = 0
-	arr[3, 4] = 0
-	arr[3, 5] = 0
-	arr[3, 6] = 0
-	arr[3, 7] = 0
-	arr[3, 8] = 0
-	arr[3, 9] = -kr[3]
-	arr[3, 10] = kf[3] * y[11]
-	arr[3, 11] = kf[3] * y[10]
-	arr[3, 12] = 0
-	arr[3, 13] = 0
-	arr[3, 14] = 0
-	arr[3, 15] = 0
-	arr[3, 16] = 0
-	arr[3, 17] = 0
-	arr[3, 18] = 0
-	arr[3, 19] = 0
-	arr[3, 20] = 0
-	arr[3, 21] = 0
-	arr[3, 22] = 0
-	arr[3, 23] = 0
-	arr[3, 24] = 0
-	arr[3, 25] = 0
-	arr[3, 26] = 0
-	arr[3, 27] = 0
-	arr[3, 28] = 0
-	arr[3, 29] = 0
-	arr[3, 30] = 0
-	arr[3, 31] = 0
-	arr[3, 32] = 0
-	arr[3, 33] = 0
-	arr[3, 34] = 0
-	arr[3, 35] = 0
-	arr[3, 36] = 0
-	arr[3, 37] = 0
-	arr[3, 38] = 0
-	arr[3, 39] = 0
-	arr[3, 40] = 0
-	arr[3, 41] = 0
-	arr[3, 42] = 0
-	arr[3, 43] = 0
-	arr[3, 44] = 0
-	arr[3, 45] = 0
-	arr[3, 46] = 0
-	arr[3, 47] = 0
-	arr[3, 48] = 0
-	arr[3, 49] = 0
-	arr[3, 50] = 0
-	arr[3, 51] = 0
-	arr[3, 52] = 0
-	arr[3, 53] = 0
-	arr[3, 54] = 0
-	arr[3, 55] = 0
-	arr[3, 56] = 0
-	arr[3, 57] = 0
-	arr[3, 58] = 0
-	arr[3, 59] = 0
-	arr[3, 60] = 0
-	arr[3, 61] = 0
-	arr[3, 62] = 0
-	arr[3, 63] = 0
-	arr[3, 64] = 0
-	arr[3, 65] = 0
-	arr[3, 66] = 0
-	arr[3, 67] = 0
-	arr[3, 68] = 0
-	arr[3, 69] = 0
-	arr[3, 70] = 0
-	arr[3, 71] = 0
-	arr[3, 72] = 0
-	arr[3, 73] = 0
-	arr[3, 74] = 0
-	arr[3, 75] = 0
-	arr[3, 76] = 0
-	arr[3, 77] = 0
-	arr[3, 78] = 0
-	arr[3, 79] = 0
-	arr[3, 80] = 0
-	arr[3, 81] = 0
-	arr[3, 82] = 0
-	arr[3, 83] = 0
-	arr[3, 84] = 0
-	arr[3, 85] = 0
-	arr[3, 86] = 0
-	arr[3, 87] = 0
-	arr[3, 88] = 0
-	arr[3, 89] = 0
-	arr[3, 90] = 0
-	arr[3, 91] = 0
-	arr[3, 92] = 0
-	arr[3, 93] = 0
-	arr[3, 94] = 0
-	arr[3, 95] = 0
-	arr[3, 96] = 0
-	arr[3, 97] = 0
-	arr[3, 98] = 0
-	arr[3, 99] = 0
-	arr[3, 100] = 0
-	arr[3, 101] = 0
-	arr[3, 102] = 0
-	arr[3, 103] = 0
-	arr[4, 0] = 0
-	arr[4, 1] = 0
-	arr[4, 2] = 0
-	arr[4, 3] = 0
-	arr[4, 4] = 0
-	arr[4, 5] = 0
-	arr[4, 6] = 0
-	arr[4, 7] = 0
-	arr[4, 8] = 0
-	arr[4, 9] = 0
-	arr[4, 10] = kf[4] * y[13]
-	arr[4, 11] = 0
-	arr[4, 12] = -kr[4]
-	arr[4, 13] = kf[4] * y[10]
-	arr[4, 14] = 0
-	arr[4, 15] = 0
-	arr[4, 16] = 0
-	arr[4, 17] = 0
-	arr[4, 18] = 0
-	arr[4, 19] = 0
-	arr[4, 20] = 0
-	arr[4, 21] = 0
-	arr[4, 22] = 0
-	arr[4, 23] = 0
-	arr[4, 24] = 0
-	arr[4, 25] = 0
-	arr[4, 26] = 0
-	arr[4, 27] = 0
-	arr[4, 28] = 0
-	arr[4, 29] = 0
-	arr[4, 30] = 0
-	arr[4, 31] = 0
-	arr[4, 32] = 0
-	arr[4, 33] = 0
-	arr[4, 34] = 0
-	arr[4, 35] = 0
-	arr[4, 36] = 0
-	arr[4, 37] = 0
-	arr[4, 38] = 0
-	arr[4, 39] = 0
-	arr[4, 40] = 0
-	arr[4, 41] = 0
-	arr[4, 42] = 0
-	arr[4, 43] = 0
-	arr[4, 44] = 0
-	arr[4, 45] = 0
-	arr[4, 46] = 0
-	arr[4, 47] = 0
-	arr[4, 48] = 0
-	arr[4, 49] = 0
-	arr[4, 50] = 0
-	arr[4, 51] = 0
-	arr[4, 52] = 0
-	arr[4, 53] = 0
-	arr[4, 54] = 0
-	arr[4, 55] = 0
-	arr[4, 56] = 0
-	arr[4, 57] = 0
-	arr[4, 58] = 0
-	arr[4, 59] = 0
-	arr[4, 60] = 0
-	arr[4, 61] = 0
-	arr[4, 62] = 0
-	arr[4, 63] = 0
-	arr[4, 64] = 0
-	arr[4, 65] = 0
-	arr[4, 66] = 0
-	arr[4, 67] = 0
-	arr[4, 68] = 0
-	arr[4, 69] = 0
-	arr[4, 70] = 0
-	arr[4, 71] = 0
-	arr[4, 72] = 0
-	arr[4, 73] = 0
-	arr[4, 74] = 0
-	arr[4, 75] = 0
-	arr[4, 76] = 0
-	arr[4, 77] = 0
-	arr[4, 78] = 0
-	arr[4, 79] = 0
-	arr[4, 80] = 0
-	arr[4, 81] = 0
-	arr[4, 82] = 0
-	arr[4, 83] = 0
-	arr[4, 84] = 0
-	arr[4, 85] = 0
-	arr[4, 86] = 0
-	arr[4, 87] = 0
-	arr[4, 88] = 0
-	arr[4, 89] = 0
-	arr[4, 90] = 0
-	arr[4, 91] = 0
-	arr[4, 92] = 0
-	arr[4, 93] = 0
-	arr[4, 94] = 0
-	arr[4, 95] = 0
-	arr[4, 96] = 0
-	arr[4, 97] = 0
-	arr[4, 98] = 0
-	arr[4, 99] = 0
-	arr[4, 100] = 0
-	arr[4, 101] = 0
-	arr[4, 102] = 0
-	arr[4, 103] = 0
-	arr[5, 0] = 0
-	arr[5, 1] = 0
-	arr[5, 2] = 0
-	arr[5, 3] = 0
-	arr[5, 4] = 0
-	arr[5, 5] = 0
-	arr[5, 6] = 0
-	arr[5, 7] = 0
-	arr[5, 8] = 0
-	arr[5, 9] = 0
-	arr[5, 10] = 0
-	arr[5, 11] = 0
-	arr[5, 12] = 0
-	arr[5, 13] = 0
-	arr[5, 14] = -kr[5]
-	arr[5, 15] = kf[5] * y[16]
-	arr[5, 16] = kf[5] * y[15]
-	arr[5, 17] = 0
-	arr[5, 18] = 0
-	arr[5, 19] = 0
-	arr[5, 20] = 0
-	arr[5, 21] = 0
-	arr[5, 22] = 0
-	arr[5, 23] = 0
-	arr[5, 24] = 0
-	arr[5, 25] = 0
-	arr[5, 26] = 0
-	arr[5, 27] = 0
-	arr[5, 28] = 0
-	arr[5, 29] = 0
-	arr[5, 30] = 0
-	arr[5, 31] = 0
-	arr[5, 32] = 0
-	arr[5, 33] = 0
-	arr[5, 34] = 0
-	arr[5, 35] = 0
-	arr[5, 36] = 0
-	arr[5, 37] = 0
-	arr[5, 38] = 0
-	arr[5, 39] = 0
-	arr[5, 40] = 0
-	arr[5, 41] = 0
-	arr[5, 42] = 0
-	arr[5, 43] = 0
-	arr[5, 44] = 0
-	arr[5, 45] = 0
-	arr[5, 46] = 0
-	arr[5, 47] = 0
-	arr[5, 48] = 0
-	arr[5, 49] = 0
-	arr[5, 50] = 0
-	arr[5, 51] = 0
-	arr[5, 52] = 0
-	arr[5, 53] = 0
-	arr[5, 54] = 0
-	arr[5, 55] = 0
-	arr[5, 56] = 0
-	arr[5, 57] = 0
-	arr[5, 58] = 0
-	arr[5, 59] = 0
-	arr[5, 60] = 0
-	arr[5, 61] = 0
-	arr[5, 62] = 0
-	arr[5, 63] = 0
-	arr[5, 64] = 0
-	arr[5, 65] = 0
-	arr[5, 66] = 0
-	arr[5, 67] = 0
-	arr[5, 68] = 0
-	arr[5, 69] = 0
-	arr[5, 70] = 0
-	arr[5, 71] = 0
-	arr[5, 72] = 0
-	arr[5, 73] = 0
-	arr[5, 74] = 0
-	arr[5, 75] = 0
-	arr[5, 76] = 0
-	arr[5, 77] = 0
-	arr[5, 78] = 0
-	arr[5, 79] = 0
-	arr[5, 80] = 0
-	arr[5, 81] = 0
-	arr[5, 82] = 0
-	arr[5, 83] = 0
-	arr[5, 84] = 0
-	arr[5, 85] = 0
-	arr[5, 86] = 0
-	arr[5, 87] = 0
-	arr[5, 88] = 0
-	arr[5, 89] = 0
-	arr[5, 90] = 0
-	arr[5, 91] = 0
-	arr[5, 92] = 0
-	arr[5, 93] = 0
-	arr[5, 94] = 0
-	arr[5, 95] = 0
-	arr[5, 96] = 0
-	arr[5, 97] = 0
-	arr[5, 98] = 0
-	arr[5, 99] = 0
-	arr[5, 100] = 0
-	arr[5, 101] = 0
-	arr[5, 102] = 0
-	arr[5, 103] = 0
-	arr[6, 0] = 0
-	arr[6, 1] = 0
-	arr[6, 2] = 0
-	arr[6, 3] = 0
-	arr[6, 4] = 0
-	arr[6, 5] = 0
-	arr[6, 6] = 0
-	arr[6, 7] = 0
-	arr[6, 8] = 0
-	arr[6, 9] = 0
-	arr[6, 10] = 0
-	arr[6, 11] = 0
-	arr[6, 12] = 0
-	arr[6, 13] = 0
-	arr[6, 14] = 0
-	arr[6, 15] = 0
-	arr[6, 16] = 0
-	arr[6, 17] = -kr[6]
-	arr[6, 18] = kf[6] * y[19]
-	arr[6, 19] = kf[6] * y[18]
-	arr[6, 20] = 0
-	arr[6, 21] = 0
-	arr[6, 22] = 0
-	arr[6, 23] = 0
-	arr[6, 24] = 0
-	arr[6, 25] = 0
-	arr[6, 26] = 0
-	arr[6, 27] = 0
-	arr[6, 28] = 0
-	arr[6, 29] = 0
-	arr[6, 30] = 0
-	arr[6, 31] = 0
-	arr[6, 32] = 0
-	arr[6, 33] = 0
-	arr[6, 34] = 0
-	arr[6, 35] = 0
-	arr[6, 36] = 0
-	arr[6, 37] = 0
-	arr[6, 38] = 0
-	arr[6, 39] = 0
-	arr[6, 40] = 0
-	arr[6, 41] = 0
-	arr[6, 42] = 0
-	arr[6, 43] = 0
-	arr[6, 44] = 0
-	arr[6, 45] = 0
-	arr[6, 46] = 0
-	arr[6, 47] = 0
-	arr[6, 48] = 0
-	arr[6, 49] = 0
-	arr[6, 50] = 0
-	arr[6, 51] = 0
-	arr[6, 52] = 0
-	arr[6, 53] = 0
-	arr[6, 54] = 0
-	arr[6, 55] = 0
-	arr[6, 56] = 0
-	arr[6, 57] = 0
-	arr[6, 58] = 0
-	arr[6, 59] = 0
-	arr[6, 60] = 0
-	arr[6, 61] = 0
-	arr[6, 62] = 0
-	arr[6, 63] = 0
-	arr[6, 64] = 0
-	arr[6, 65] = 0
-	arr[6, 66] = 0
-	arr[6, 67] = 0
-	arr[6, 68] = 0
-	arr[6, 69] = 0
-	arr[6, 70] = 0
-	arr[6, 71] = 0
-	arr[6, 72] = 0
-	arr[6, 73] = 0
-	arr[6, 74] = 0
-	arr[6, 75] = 0
-	arr[6, 76] = 0
-	arr[6, 77] = 0
-	arr[6, 78] = 0
-	arr[6, 79] = 0
-	arr[6, 80] = 0
-	arr[6, 81] = 0
-	arr[6, 82] = 0
-	arr[6, 83] = 0
-	arr[6, 84] = 0
-	arr[6, 85] = 0
-	arr[6, 86] = 0
-	arr[6, 87] = 0
-	arr[6, 88] = 0
-	arr[6, 89] = 0
-	arr[6, 90] = 0
-	arr[6, 91] = 0
-	arr[6, 92] = 0
-	arr[6, 93] = 0
-	arr[6, 94] = 0
-	arr[6, 95] = 0
-	arr[6, 96] = 0
-	arr[6, 97] = 0
-	arr[6, 98] = 0
-	arr[6, 99] = 0
-	arr[6, 100] = 0
-	arr[6, 101] = 0
-	arr[6, 102] = 0
-	arr[6, 103] = 0
-	arr[7, 0] = 0
-	arr[7, 1] = 0
-	arr[7, 2] = 0
-	arr[7, 3] = 0
-	arr[7, 4] = 0
-	arr[7, 5] = 0
-	arr[7, 6] = 0
-	arr[7, 7] = 0
-	arr[7, 8] = 0
-	arr[7, 9] = 0
-	arr[7, 10] = 0
-	arr[7, 11] = 0
-	arr[7, 12] = 0
-	arr[7, 13] = 0
-	arr[7, 14] = 0
-	arr[7, 15] = 0
-	arr[7, 16] = 0
-	arr[7, 17] = 0
-	arr[7, 18] = 0
-	arr[7, 19] = 0
-	arr[7, 20] = -kr[7]
-	arr[7, 21] = kf[7] * y[22]
-	arr[7, 22] = kf[7] * y[21]
-	arr[7, 23] = 0
-	arr[7, 24] = 0
-	arr[7, 25] = 0
-	arr[7, 26] = 0
-	arr[7, 27] = 0
-	arr[7, 28] = 0
-	arr[7, 29] = 0
-	arr[7, 30] = 0
-	arr[7, 31] = 0
-	arr[7, 32] = 0
-	arr[7, 33] = 0
-	arr[7, 34] = 0
-	arr[7, 35] = 0
-	arr[7, 36] = 0
-	arr[7, 37] = 0
-	arr[7, 38] = 0
-	arr[7, 39] = 0
-	arr[7, 40] = 0
-	arr[7, 41] = 0
-	arr[7, 42] = 0
-	arr[7, 43] = 0
-	arr[7, 44] = 0
-	arr[7, 45] = 0
-	arr[7, 46] = 0
-	arr[7, 47] = 0
-	arr[7, 48] = 0
-	arr[7, 49] = 0
-	arr[7, 50] = 0
-	arr[7, 51] = 0
-	arr[7, 52] = 0
-	arr[7, 53] = 0
-	arr[7, 54] = 0
-	arr[7, 55] = 0
-	arr[7, 56] = 0
-	arr[7, 57] = 0
-	arr[7, 58] = 0
-	arr[7, 59] = 0
-	arr[7, 60] = 0
-	arr[7, 61] = 0
-	arr[7, 62] = 0
-	arr[7, 63] = 0
-	arr[7, 64] = 0
-	arr[7, 65] = 0
-	arr[7, 66] = 0
-	arr[7, 67] = 0
-	arr[7, 68] = 0
-	arr[7, 69] = 0
-	arr[7, 70] = 0
-	arr[7, 71] = 0
-	arr[7, 72] = 0
-	arr[7, 73] = 0
-	arr[7, 74] = 0
-	arr[7, 75] = 0
-	arr[7, 76] = 0
-	arr[7, 77] = 0
-	arr[7, 78] = 0
-	arr[7, 79] = 0
-	arr[7, 80] = 0
-	arr[7, 81] = 0
-	arr[7, 82] = 0
-	arr[7, 83] = 0
-	arr[7, 84] = 0
-	arr[7, 85] = 0
-	arr[7, 86] = 0
-	arr[7, 87] = 0
-	arr[7, 88] = 0
-	arr[7, 89] = 0
-	arr[7, 90] = 0
-	arr[7, 91] = 0
-	arr[7, 92] = 0
-	arr[7, 93] = 0
-	arr[7, 94] = 0
-	arr[7, 95] = 0
-	arr[7, 96] = 0
-	arr[7, 97] = 0
-	arr[7, 98] = 0
-	arr[7, 99] = 0
-	arr[7, 100] = 0
-	arr[7, 101] = 0
-	arr[7, 102] = 0
-	arr[7, 103] = 0
-	arr[8, 0] = 0
-	arr[8, 1] = 0
-	arr[8, 2] = 0
-	arr[8, 3] = 0
-	arr[8, 4] = 0
-	arr[8, 5] = 0
-	arr[8, 6] = 0
-	arr[8, 7] = 0
-	arr[8, 8] = 0
-	arr[8, 9] = 0
-	arr[8, 10] = 0
-	arr[8, 11] = 0
-	arr[8, 12] = 0
-	arr[8, 13] = 0
-	arr[8, 14] = 0
-	arr[8, 15] = 0
-	arr[8, 16] = 0
-	arr[8, 17] = 0
-	arr[8, 18] = 0
-	arr[8, 19] = 0
-	arr[8, 20] = 0
-	arr[8, 21] = 0
-	arr[8, 22] = 0
-	arr[8, 23] = -kr[8]
-	arr[8, 24] = kf[8] * y[25]
-	arr[8, 25] = kf[8] * y[24]
-	arr[8, 26] = 0
-	arr[8, 27] = 0
-	arr[8, 28] = 0
-	arr[8, 29] = 0
-	arr[8, 30] = 0
-	arr[8, 31] = 0
-	arr[8, 32] = 0
-	arr[8, 33] = 0
-	arr[8, 34] = 0
-	arr[8, 35] = 0
-	arr[8, 36] = 0
-	arr[8, 37] = 0
-	arr[8, 38] = 0
-	arr[8, 39] = 0
-	arr[8, 40] = 0
-	arr[8, 41] = 0
-	arr[8, 42] = 0
-	arr[8, 43] = 0
-	arr[8, 44] = 0
-	arr[8, 45] = 0
-	arr[8, 46] = 0
-	arr[8, 47] = 0
-	arr[8, 48] = 0
-	arr[8, 49] = 0
-	arr[8, 50] = 0
-	arr[8, 51] = 0
-	arr[8, 52] = 0
-	arr[8, 53] = 0
-	arr[8, 54] = 0
-	arr[8, 55] = 0
-	arr[8, 56] = 0
-	arr[8, 57] = 0
-	arr[8, 58] = 0
-	arr[8, 59] = 0
-	arr[8, 60] = 0
-	arr[8, 61] = 0
-	arr[8, 62] = 0
-	arr[8, 63] = 0
-	arr[8, 64] = 0
-	arr[8, 65] = 0
-	arr[8, 66] = 0
-	arr[8, 67] = 0
-	arr[8, 68] = 0
-	arr[8, 69] = 0
-	arr[8, 70] = 0
-	arr[8, 71] = 0
-	arr[8, 72] = 0
-	arr[8, 73] = 0
-	arr[8, 74] = 0
-	arr[8, 75] = 0
-	arr[8, 76] = 0
-	arr[8, 77] = 0
-	arr[8, 78] = 0
-	arr[8, 79] = 0
-	arr[8, 80] = 0
-	arr[8, 81] = 0
-	arr[8, 82] = 0
-	arr[8, 83] = 0
-	arr[8, 84] = 0
-	arr[8, 85] = 0
-	arr[8, 86] = 0
-	arr[8, 87] = 0
-	arr[8, 88] = 0
-	arr[8, 89] = 0
-	arr[8, 90] = 0
-	arr[8, 91] = 0
-	arr[8, 92] = 0
-	arr[8, 93] = 0
-	arr[8, 94] = 0
-	arr[8, 95] = 0
-	arr[8, 96] = 0
-	arr[8, 97] = 0
-	arr[8, 98] = 0
-	arr[8, 99] = 0
-	arr[8, 100] = 0
-	arr[8, 101] = 0
-	arr[8, 102] = 0
-	arr[8, 103] = 0
-	arr[9, 0] = 0
-	arr[9, 1] = 0
-	arr[9, 2] = 0
-	arr[9, 3] = 0
-	arr[9, 4] = 0
-	arr[9, 5] = 0
-	arr[9, 6] = 0
-	arr[9, 7] = 0
-	arr[9, 8] = 0
-	arr[9, 9] = 0
-	arr[9, 10] = 0
-	arr[9, 11] = 0
-	arr[9, 12] = 0
-	arr[9, 13] = 0
-	arr[9, 14] = 0
-	arr[9, 15] = 0
-	arr[9, 16] = 0
-	arr[9, 17] = 0
-	arr[9, 18] = 0
-	arr[9, 19] = 0
-	arr[9, 20] = 0
-	arr[9, 21] = 0
-	arr[9, 22] = 0
-	arr[9, 23] = 0
-	arr[9, 24] = 0
-	arr[9, 25] = 0
-	arr[9, 26] = -kr[9]
-	arr[9, 27] = kf[9] * y[28]
-	arr[9, 28] = kf[9] * y[27]
-	arr[9, 29] = 0
-	arr[9, 30] = 0
-	arr[9, 31] = 0
-	arr[9, 32] = 0
-	arr[9, 33] = 0
-	arr[9, 34] = 0
-	arr[9, 35] = 0
-	arr[9, 36] = 0
-	arr[9, 37] = 0
-	arr[9, 38] = 0
-	arr[9, 39] = 0
-	arr[9, 40] = 0
-	arr[9, 41] = 0
-	arr[9, 42] = 0
-	arr[9, 43] = 0
-	arr[9, 44] = 0
-	arr[9, 45] = 0
-	arr[9, 46] = 0
-	arr[9, 47] = 0
-	arr[9, 48] = 0
-	arr[9, 49] = 0
-	arr[9, 50] = 0
-	arr[9, 51] = 0
-	arr[9, 52] = 0
-	arr[9, 53] = 0
-	arr[9, 54] = 0
-	arr[9, 55] = 0
-	arr[9, 56] = 0
-	arr[9, 57] = 0
-	arr[9, 58] = 0
-	arr[9, 59] = 0
-	arr[9, 60] = 0
-	arr[9, 61] = 0
-	arr[9, 62] = 0
-	arr[9, 63] = 0
-	arr[9, 64] = 0
-	arr[9, 65] = 0
-	arr[9, 66] = 0
-	arr[9, 67] = 0
-	arr[9, 68] = 0
-	arr[9, 69] = 0
-	arr[9, 70] = 0
-	arr[9, 71] = 0
-	arr[9, 72] = 0
-	arr[9, 73] = 0
-	arr[9, 74] = 0
-	arr[9, 75] = 0
-	arr[9, 76] = 0
-	arr[9, 77] = 0
-	arr[9, 78] = 0
-	arr[9, 79] = 0
-	arr[9, 80] = 0
-	arr[9, 81] = 0
-	arr[9, 82] = 0
-	arr[9, 83] = 0
-	arr[9, 84] = 0
-	arr[9, 85] = 0
-	arr[9, 86] = 0
-	arr[9, 87] = 0
-	arr[9, 88] = 0
-	arr[9, 89] = 0
-	arr[9, 90] = 0
-	arr[9, 91] = 0
-	arr[9, 92] = 0
-	arr[9, 93] = 0
-	arr[9, 94] = 0
-	arr[9, 95] = 0
-	arr[9, 96] = 0
-	arr[9, 97] = 0
-	arr[9, 98] = 0
-	arr[9, 99] = 0
-	arr[9, 100] = 0
-	arr[9, 101] = 0
-	arr[9, 102] = 0
-	arr[9, 103] = 0
-	arr[10, 0] = 0
-	arr[10, 1] = 0
-	arr[10, 2] = 0
-	arr[10, 3] = 0
-	arr[10, 4] = 0
-	arr[10, 5] = 0
-	arr[10, 6] = 0
-	arr[10, 7] = 0
-	arr[10, 8] = 0
-	arr[10, 9] = 0
-	arr[10, 10] = 0
-	arr[10, 11] = 0
-	arr[10, 12] = 0
-	arr[10, 13] = 0
-	arr[10, 14] = 0
-	arr[10, 15] = 0
-	arr[10, 16] = 0
-	arr[10, 17] = 0
-	arr[10, 18] = 0
-	arr[10, 19] = 0
-	arr[10, 20] = 0
-	arr[10, 21] = 0
-	arr[10, 22] = 0
-	arr[10, 23] = 0
-	arr[10, 24] = 0
-	arr[10, 25] = 0
-	arr[10, 26] = 0
-	arr[10, 27] = 0
-	arr[10, 28] = 0
-	arr[10, 29] = -kr[10]
-	arr[10, 30] = kf[10] * y[31]
-	arr[10, 31] = kf[10] * y[30]
-	arr[10, 32] = 0
-	arr[10, 33] = 0
-	arr[10, 34] = 0
-	arr[10, 35] = 0
-	arr[10, 36] = 0
-	arr[10, 37] = 0
-	arr[10, 38] = 0
-	arr[10, 39] = 0
-	arr[10, 40] = 0
-	arr[10, 41] = 0
-	arr[10, 42] = 0
-	arr[10, 43] = 0
-	arr[10, 44] = 0
-	arr[10, 45] = 0
-	arr[10, 46] = 0
-	arr[10, 47] = 0
-	arr[10, 48] = 0
-	arr[10, 49] = 0
-	arr[10, 50] = 0
-	arr[10, 51] = 0
-	arr[10, 52] = 0
-	arr[10, 53] = 0
-	arr[10, 54] = 0
-	arr[10, 55] = 0
-	arr[10, 56] = 0
-	arr[10, 57] = 0
-	arr[10, 58] = 0
-	arr[10, 59] = 0
-	arr[10, 60] = 0
-	arr[10, 61] = 0
-	arr[10, 62] = 0
-	arr[10, 63] = 0
-	arr[10, 64] = 0
-	arr[10, 65] = 0
-	arr[10, 66] = 0
-	arr[10, 67] = 0
-	arr[10, 68] = 0
-	arr[10, 69] = 0
-	arr[10, 70] = 0
-	arr[10, 71] = 0
-	arr[10, 72] = 0
-	arr[10, 73] = 0
-	arr[10, 74] = 0
-	arr[10, 75] = 0
-	arr[10, 76] = 0
-	arr[10, 77] = 0
-	arr[10, 78] = 0
-	arr[10, 79] = 0
-	arr[10, 80] = 0
-	arr[10, 81] = 0
-	arr[10, 82] = 0
-	arr[10, 83] = 0
-	arr[10, 84] = 0
-	arr[10, 85] = 0
-	arr[10, 86] = 0
-	arr[10, 87] = 0
-	arr[10, 88] = 0
-	arr[10, 89] = 0
-	arr[10, 90] = 0
-	arr[10, 91] = 0
-	arr[10, 92] = 0
-	arr[10, 93] = 0
-	arr[10, 94] = 0
-	arr[10, 95] = 0
-	arr[10, 96] = 0
-	arr[10, 97] = 0
-	arr[10, 98] = 0
-	arr[10, 99] = 0
-	arr[10, 100] = 0
-	arr[10, 101] = 0
-	arr[10, 102] = 0
-	arr[10, 103] = 0
-	arr[11, 0] = 0
-	arr[11, 1] = 0
-	arr[11, 2] = 0
-	arr[11, 3] = 0
-	arr[11, 4] = 0
-	arr[11, 5] = 0
-	arr[11, 6] = 0
-	arr[11, 7] = 0
-	arr[11, 8] = 0
-	arr[11, 9] = 0
-	arr[11, 10] = 0
-	arr[11, 11] = 0
-	arr[11, 12] = 0
-	arr[11, 13] = 0
-	arr[11, 14] = 0
-	arr[11, 15] = 0
-	arr[11, 16] = 0
-	arr[11, 17] = 0
-	arr[11, 18] = 0
-	arr[11, 19] = 0
-	arr[11, 20] = 0
-	arr[11, 21] = 0
-	arr[11, 22] = 0
-	arr[11, 23] = 0
-	arr[11, 24] = 0
-	arr[11, 25] = 0
-	arr[11, 26] = 0
-	arr[11, 27] = 0
-	arr[11, 28] = 0
-	arr[11, 29] = 0
-	arr[11, 30] = 0
-	arr[11, 31] = 0
-	arr[11, 32] = -kr[11]
-	arr[11, 33] = kf[11] * y[34]
-	arr[11, 34] = kf[11] * y[33]
-	arr[11, 35] = 0
-	arr[11, 36] = 0
-	arr[11, 37] = 0
-	arr[11, 38] = 0
-	arr[11, 39] = 0
-	arr[11, 40] = 0
-	arr[11, 41] = 0
-	arr[11, 42] = 0
-	arr[11, 43] = 0
-	arr[11, 44] = 0
-	arr[11, 45] = 0
-	arr[11, 46] = 0
-	arr[11, 47] = 0
-	arr[11, 48] = 0
-	arr[11, 49] = 0
-	arr[11, 50] = 0
-	arr[11, 51] = 0
-	arr[11, 52] = 0
-	arr[11, 53] = 0
-	arr[11, 54] = 0
-	arr[11, 55] = 0
-	arr[11, 56] = 0
-	arr[11, 57] = 0
-	arr[11, 58] = 0
-	arr[11, 59] = 0
-	arr[11, 60] = 0
-	arr[11, 61] = 0
-	arr[11, 62] = 0
-	arr[11, 63] = 0
-	arr[11, 64] = 0
-	arr[11, 65] = 0
-	arr[11, 66] = 0
-	arr[11, 67] = 0
-	arr[11, 68] = 0
-	arr[11, 69] = 0
-	arr[11, 70] = 0
-	arr[11, 71] = 0
-	arr[11, 72] = 0
-	arr[11, 73] = 0
-	arr[11, 74] = 0
-	arr[11, 75] = 0
-	arr[11, 76] = 0
-	arr[11, 77] = 0
-	arr[11, 78] = 0
-	arr[11, 79] = 0
-	arr[11, 80] = 0
-	arr[11, 81] = 0
-	arr[11, 82] = 0
-	arr[11, 83] = 0
-	arr[11, 84] = 0
-	arr[11, 85] = 0
-	arr[11, 86] = 0
-	arr[11, 87] = 0
-	arr[11, 88] = 0
-	arr[11, 89] = 0
-	arr[11, 90] = 0
-	arr[11, 91] = 0
-	arr[11, 92] = 0
-	arr[11, 93] = 0
-	arr[11, 94] = 0
-	arr[11, 95] = 0
-	arr[11, 96] = 0
-	arr[11, 97] = 0
-	arr[11, 98] = 0
-	arr[11, 99] = 0
-	arr[11, 100] = 0
-	arr[11, 101] = 0
-	arr[11, 102] = 0
-	arr[11, 103] = 0
-	arr[12, 0] = 0
-	arr[12, 1] = 0
-	arr[12, 2] = 0
-	arr[12, 3] = 0
-	arr[12, 4] = 0
-	arr[12, 5] = 0
-	arr[12, 6] = 0
-	arr[12, 7] = 0
-	arr[12, 8] = 0
-	arr[12, 9] = 0
-	arr[12, 10] = 0
-	arr[12, 11] = 0
-	arr[12, 12] = 0
-	arr[12, 13] = 0
-	arr[12, 14] = 0
-	arr[12, 15] = 0
-	arr[12, 16] = 0
-	arr[12, 17] = 0
-	arr[12, 18] = 0
-	arr[12, 19] = 0
-	arr[12, 20] = 0
-	arr[12, 21] = 0
-	arr[12, 22] = 0
-	arr[12, 23] = 0
-	arr[12, 24] = 0
-	arr[12, 25] = 0
-	arr[12, 26] = 0
-	arr[12, 27] = 0
-	arr[12, 28] = 0
-	arr[12, 29] = 0
-	arr[12, 30] = 0
-	arr[12, 31] = 0
-	arr[12, 32] = 0
-	arr[12, 33] = 0
-	arr[12, 34] = 0
-	arr[12, 35] = -kr[12]
-	arr[12, 36] = kf[12] * y[37]
-	arr[12, 37] = kf[12] * y[36]
-	arr[12, 38] = 0
-	arr[12, 39] = 0
-	arr[12, 40] = 0
-	arr[12, 41] = 0
-	arr[12, 42] = 0
-	arr[12, 43] = 0
-	arr[12, 44] = 0
-	arr[12, 45] = 0
-	arr[12, 46] = 0
-	arr[12, 47] = 0
-	arr[12, 48] = 0
-	arr[12, 49] = 0
-	arr[12, 50] = 0
-	arr[12, 51] = 0
-	arr[12, 52] = 0
-	arr[12, 53] = 0
-	arr[12, 54] = 0
-	arr[12, 55] = 0
-	arr[12, 56] = 0
-	arr[12, 57] = 0
-	arr[12, 58] = 0
-	arr[12, 59] = 0
-	arr[12, 60] = 0
-	arr[12, 61] = 0
-	arr[12, 62] = 0
-	arr[12, 63] = 0
-	arr[12, 64] = 0
-	arr[12, 65] = 0
-	arr[12, 66] = 0
-	arr[12, 67] = 0
-	arr[12, 68] = 0
-	arr[12, 69] = 0
-	arr[12, 70] = 0
-	arr[12, 71] = 0
-	arr[12, 72] = 0
-	arr[12, 73] = 0
-	arr[12, 74] = 0
-	arr[12, 75] = 0
-	arr[12, 76] = 0
-	arr[12, 77] = 0
-	arr[12, 78] = 0
-	arr[12, 79] = 0
-	arr[12, 80] = 0
-	arr[12, 81] = 0
-	arr[12, 82] = 0
-	arr[12, 83] = 0
-	arr[12, 84] = 0
-	arr[12, 85] = 0
-	arr[12, 86] = 0
-	arr[12, 87] = 0
-	arr[12, 88] = 0
-	arr[12, 89] = 0
-	arr[12, 90] = 0
-	arr[12, 91] = 0
-	arr[12, 92] = 0
-	arr[12, 93] = 0
-	arr[12, 94] = 0
-	arr[12, 95] = 0
-	arr[12, 96] = 0
-	arr[12, 97] = 0
-	arr[12, 98] = 0
-	arr[12, 99] = 0
-	arr[12, 100] = 0
-	arr[12, 101] = 0
-	arr[12, 102] = 0
-	arr[12, 103] = 0
-	arr[13, 0] = 0
-	arr[13, 1] = 0
-	arr[13, 2] = 0
-	arr[13, 3] = 0
-	arr[13, 4] = 0
-	arr[13, 5] = 0
-	arr[13, 6] = 0
-	arr[13, 7] = 0
-	arr[13, 8] = 0
-	arr[13, 9] = 0
-	arr[13, 10] = 0
-	arr[13, 11] = 0
-	arr[13, 12] = 0
-	arr[13, 13] = 0
-	arr[13, 14] = 0
-	arr[13, 15] = 0
-	arr[13, 16] = 0
-	arr[13, 17] = 0
-	arr[13, 18] = 0
-	arr[13, 19] = 0
-	arr[13, 20] = 0
-	arr[13, 21] = 0
-	arr[13, 22] = 0
-	arr[13, 23] = 0
-	arr[13, 24] = 0
-	arr[13, 25] = 0
-	arr[13, 26] = 0
-	arr[13, 27] = 0
-	arr[13, 28] = 0
-	arr[13, 29] = 0
-	arr[13, 30] = 0
-	arr[13, 31] = 0
-	arr[13, 32] = 0
-	arr[13, 33] = 0
-	arr[13, 34] = 0
-	arr[13, 35] = 0
-	arr[13, 36] = 0
-	arr[13, 37] = 0
-	arr[13, 38] = -kr[13]
-	arr[13, 39] = kf[13] * y[40]
-	arr[13, 40] = kf[13] * y[39]
-	arr[13, 41] = 0
-	arr[13, 42] = 0
-	arr[13, 43] = 0
-	arr[13, 44] = 0
-	arr[13, 45] = 0
-	arr[13, 46] = 0
-	arr[13, 47] = 0
-	arr[13, 48] = 0
-	arr[13, 49] = 0
-	arr[13, 50] = 0
-	arr[13, 51] = 0
-	arr[13, 52] = 0
-	arr[13, 53] = 0
-	arr[13, 54] = 0
-	arr[13, 55] = 0
-	arr[13, 56] = 0
-	arr[13, 57] = 0
-	arr[13, 58] = 0
-	arr[13, 59] = 0
-	arr[13, 60] = 0
-	arr[13, 61] = 0
-	arr[13, 62] = 0
-	arr[13, 63] = 0
-	arr[13, 64] = 0
-	arr[13, 65] = 0
-	arr[13, 66] = 0
-	arr[13, 67] = 0
-	arr[13, 68] = 0
-	arr[13, 69] = 0
-	arr[13, 70] = 0
-	arr[13, 71] = 0
-	arr[13, 72] = 0
-	arr[13, 73] = 0
-	arr[13, 74] = 0
-	arr[13, 75] = 0
-	arr[13, 76] = 0
-	arr[13, 77] = 0
-	arr[13, 78] = 0
-	arr[13, 79] = 0
-	arr[13, 80] = 0
-	arr[13, 81] = 0
-	arr[13, 82] = 0
-	arr[13, 83] = 0
-	arr[13, 84] = 0
-	arr[13, 85] = 0
-	arr[13, 86] = 0
-	arr[13, 87] = 0
-	arr[13, 88] = 0
-	arr[13, 89] = 0
-	arr[13, 90] = 0
-	arr[13, 91] = 0
-	arr[13, 92] = 0
-	arr[13, 93] = 0
-	arr[13, 94] = 0
-	arr[13, 95] = 0
-	arr[13, 96] = 0
-	arr[13, 97] = 0
-	arr[13, 98] = 0
-	arr[13, 99] = 0
-	arr[13, 100] = 0
-	arr[13, 101] = 0
-	arr[13, 102] = 0
-	arr[13, 103] = 0
-	arr[14, 0] = 0
-	arr[14, 1] = 0
-	arr[14, 2] = 0
-	arr[14, 3] = 0
-	arr[14, 4] = 0
-	arr[14, 5] = 0
-	arr[14, 6] = 0
-	arr[14, 7] = 0
-	arr[14, 8] = 0
-	arr[14, 9] = 0
-	arr[14, 10] = 0
-	arr[14, 11] = 0
-	arr[14, 12] = 0
-	arr[14, 13] = 0
-	arr[14, 14] = 0
-	arr[14, 15] = 0
-	arr[14, 16] = 0
-	arr[14, 17] = 0
-	arr[14, 18] = 0
-	arr[14, 19] = 0
-	arr[14, 20] = 0
-	arr[14, 21] = 0
-	arr[14, 22] = kf[14] * y[42]
-	arr[14, 23] = 0
-	arr[14, 24] = 0
-	arr[14, 25] = 0
-	arr[14, 26] = 0
-	arr[14, 27] = 0
-	arr[14, 28] = 0
-	arr[14, 29] = 0
-	arr[14, 30] = 0
-	arr[14, 31] = 0
-	arr[14, 32] = 0
-	arr[14, 33] = 0
-	arr[14, 34] = 0
-	arr[14, 35] = 0
-	arr[14, 36] = 0
-	arr[14, 37] = 0
-	arr[14, 38] = 0
-	arr[14, 39] = 0
-	arr[14, 40] = 0
-	arr[14, 41] = -kr[14]
-	arr[14, 42] = kf[14] * y[22]
-	arr[14, 43] = 0
-	arr[14, 44] = 0
-	arr[14, 45] = 0
-	arr[14, 46] = 0
-	arr[14, 47] = 0
-	arr[14, 48] = 0
-	arr[14, 49] = 0
-	arr[14, 50] = 0
-	arr[14, 51] = 0
-	arr[14, 52] = 0
-	arr[14, 53] = 0
-	arr[14, 54] = 0
-	arr[14, 55] = 0
-	arr[14, 56] = 0
-	arr[14, 57] = 0
-	arr[14, 58] = 0
-	arr[14, 59] = 0
-	arr[14, 60] = 0
-	arr[14, 61] = 0
-	arr[14, 62] = 0
-	arr[14, 63] = 0
-	arr[14, 64] = 0
-	arr[14, 65] = 0
-	arr[14, 66] = 0
-	arr[14, 67] = 0
-	arr[14, 68] = 0
-	arr[14, 69] = 0
-	arr[14, 70] = 0
-	arr[14, 71] = 0
-	arr[14, 72] = 0
-	arr[14, 73] = 0
-	arr[14, 74] = 0
-	arr[14, 75] = 0
-	arr[14, 76] = 0
-	arr[14, 77] = 0
-	arr[14, 78] = 0
-	arr[14, 79] = 0
-	arr[14, 80] = 0
-	arr[14, 81] = 0
-	arr[14, 82] = 0
-	arr[14, 83] = 0
-	arr[14, 84] = 0
-	arr[14, 85] = 0
-	arr[14, 86] = 0
-	arr[14, 87] = 0
-	arr[14, 88] = 0
-	arr[14, 89] = 0
-	arr[14, 90] = 0
-	arr[14, 91] = 0
-	arr[14, 92] = 0
-	arr[14, 93] = 0
-	arr[14, 94] = 0
-	arr[14, 95] = 0
-	arr[14, 96] = 0
-	arr[14, 97] = 0
-	arr[14, 98] = 0
-	arr[14, 99] = 0
-	arr[14, 100] = 0
-	arr[14, 101] = 0
-	arr[14, 102] = 0
-	arr[14, 103] = 0
-	arr[15, 0] = 0
-	arr[15, 1] = 0
-	arr[15, 2] = 0
-	arr[15, 3] = 0
-	arr[15, 4] = 0
-	arr[15, 5] = 0
-	arr[15, 6] = 0
-	arr[15, 7] = 0
-	arr[15, 8] = 0
-	arr[15, 9] = 0
-	arr[15, 10] = kf[15] * y[44]
-	arr[15, 11] = 0
-	arr[15, 12] = 0
-	arr[15, 13] = 0
-	arr[15, 14] = 0
-	arr[15, 15] = 0
-	arr[15, 16] = 0
-	arr[15, 17] = 0
-	arr[15, 18] = 0
-	arr[15, 19] = 0
-	arr[15, 20] = 0
-	arr[15, 21] = 0
-	arr[15, 22] = 0
-	arr[15, 23] = 0
-	arr[15, 24] = 0
-	arr[15, 25] = 0
-	arr[15, 26] = 0
-	arr[15, 27] = 0
-	arr[15, 28] = 0
-	arr[15, 29] = 0
-	arr[15, 30] = 0
-	arr[15, 31] = 0
-	arr[15, 32] = 0
-	arr[15, 33] = 0
-	arr[15, 34] = 0
-	arr[15, 35] = 0
-	arr[15, 36] = 0
-	arr[15, 37] = 0
-	arr[15, 38] = 0
-	arr[15, 39] = 0
-	arr[15, 40] = 0
-	arr[15, 41] = 0
-	arr[15, 42] = 0
-	arr[15, 43] = -kr[15]
-	arr[15, 44] = kf[15] * y[10]
-	arr[15, 45] = 0
-	arr[15, 46] = 0
-	arr[15, 47] = 0
-	arr[15, 48] = 0
-	arr[15, 49] = 0
-	arr[15, 50] = 0
-	arr[15, 51] = 0
-	arr[15, 52] = 0
-	arr[15, 53] = 0
-	arr[15, 54] = 0
-	arr[15, 55] = 0
-	arr[15, 56] = 0
-	arr[15, 57] = 0
-	arr[15, 58] = 0
-	arr[15, 59] = 0
-	arr[15, 60] = 0
-	arr[15, 61] = 0
-	arr[15, 62] = 0
-	arr[15, 63] = 0
-	arr[15, 64] = 0
-	arr[15, 65] = 0
-	arr[15, 66] = 0
-	arr[15, 67] = 0
-	arr[15, 68] = 0
-	arr[15, 69] = 0
-	arr[15, 70] = 0
-	arr[15, 71] = 0
-	arr[15, 72] = 0
-	arr[15, 73] = 0
-	arr[15, 74] = 0
-	arr[15, 75] = 0
-	arr[15, 76] = 0
-	arr[15, 77] = 0
-	arr[15, 78] = 0
-	arr[15, 79] = 0
-	arr[15, 80] = 0
-	arr[15, 81] = 0
-	arr[15, 82] = 0
-	arr[15, 83] = 0
-	arr[15, 84] = 0
-	arr[15, 85] = 0
-	arr[15, 86] = 0
-	arr[15, 87] = 0
-	arr[15, 88] = 0
-	arr[15, 89] = 0
-	arr[15, 90] = 0
-	arr[15, 91] = 0
-	arr[15, 92] = 0
-	arr[15, 93] = 0
-	arr[15, 94] = 0
-	arr[15, 95] = 0
-	arr[15, 96] = 0
-	arr[15, 97] = 0
-	arr[15, 98] = 0
-	arr[15, 99] = 0
-	arr[15, 100] = 0
-	arr[15, 101] = 0
-	arr[15, 102] = 0
-	arr[15, 103] = 0
-	arr[16, 0] = 0
-	arr[16, 1] = 0
-	arr[16, 2] = 0
-	arr[16, 3] = 0
-	arr[16, 4] = 0
-	arr[16, 5] = 0
-	arr[16, 6] = 0
-	arr[16, 7] = 0
-	arr[16, 8] = 0
-	arr[16, 9] = 0
-	arr[16, 10] = 0
-	arr[16, 11] = 0
-	arr[16, 12] = 0
-	arr[16, 13] = 0
-	arr[16, 14] = 0
-	arr[16, 15] = 0
-	arr[16, 16] = 0
-	arr[16, 17] = 0
-	arr[16, 18] = 0
-	arr[16, 19] = 0
-	arr[16, 20] = 0
-	arr[16, 21] = 0
-	arr[16, 22] = 0
-	arr[16, 23] = 0
-	arr[16, 24] = 0
-	arr[16, 25] = 0
-	arr[16, 26] = 0
-	arr[16, 27] = 0
-	arr[16, 28] = 0
-	arr[16, 29] = 0
-	arr[16, 30] = 0
-	arr[16, 31] = 0
-	arr[16, 32] = 0
-	arr[16, 33] = 0
-	arr[16, 34] = 0
-	arr[16, 35] = 0
-	arr[16, 36] = 0
-	arr[16, 37] = 0
-	arr[16, 38] = 0
-	arr[16, 39] = 0
-	arr[16, 40] = 0
-	arr[16, 41] = 0
-	arr[16, 42] = 0
-	arr[16, 43] = 0
-	arr[16, 44] = 0
-	arr[16, 45] = -kr[16]
-	arr[16, 46] = kf[16] * y[47]
-	arr[16, 47] = kf[16] * y[46]
-	arr[16, 48] = 0
-	arr[16, 49] = 0
-	arr[16, 50] = 0
-	arr[16, 51] = 0
-	arr[16, 52] = 0
-	arr[16, 53] = 0
-	arr[16, 54] = 0
-	arr[16, 55] = 0
-	arr[16, 56] = 0
-	arr[16, 57] = 0
-	arr[16, 58] = 0
-	arr[16, 59] = 0
-	arr[16, 60] = 0
-	arr[16, 61] = 0
-	arr[16, 62] = 0
-	arr[16, 63] = 0
-	arr[16, 64] = 0
-	arr[16, 65] = 0
-	arr[16, 66] = 0
-	arr[16, 67] = 0
-	arr[16, 68] = 0
-	arr[16, 69] = 0
-	arr[16, 70] = 0
-	arr[16, 71] = 0
-	arr[16, 72] = 0
-	arr[16, 73] = 0
-	arr[16, 74] = 0
-	arr[16, 75] = 0
-	arr[16, 76] = 0
-	arr[16, 77] = 0
-	arr[16, 78] = 0
-	arr[16, 79] = 0
-	arr[16, 80] = 0
-	arr[16, 81] = 0
-	arr[16, 82] = 0
-	arr[16, 83] = 0
-	arr[16, 84] = 0
-	arr[16, 85] = 0
-	arr[16, 86] = 0
-	arr[16, 87] = 0
-	arr[16, 88] = 0
-	arr[16, 89] = 0
-	arr[16, 90] = 0
-	arr[16, 91] = 0
-	arr[16, 92] = 0
-	arr[16, 93] = 0
-	arr[16, 94] = 0
-	arr[16, 95] = 0
-	arr[16, 96] = 0
-	arr[16, 97] = 0
-	arr[16, 98] = 0
-	arr[16, 99] = 0
-	arr[16, 100] = 0
-	arr[16, 101] = 0
-	arr[16, 102] = 0
-	arr[16, 103] = 0
-	arr[17, 0] = 0
-	arr[17, 1] = 0
-	arr[17, 2] = 0
-	arr[17, 3] = 0
-	arr[17, 4] = 0
-	arr[17, 5] = 0
-	arr[17, 6] = 0
-	arr[17, 7] = 0
-	arr[17, 8] = 0
-	arr[17, 9] = 0
-	arr[17, 10] = 0
-	arr[17, 11] = 0
-	arr[17, 12] = 0
-	arr[17, 13] = 0
-	arr[17, 14] = 0
-	arr[17, 15] = 0
-	arr[17, 16] = 0
-	arr[17, 17] = 0
-	arr[17, 18] = 0
-	arr[17, 19] = 0
-	arr[17, 20] = 0
-	arr[17, 21] = 0
-	arr[17, 22] = 0
-	arr[17, 23] = 0
-	arr[17, 24] = 0
-	arr[17, 25] = 0
-	arr[17, 26] = 0
-	arr[17, 27] = 0
-	arr[17, 28] = 0
-	arr[17, 29] = 0
-	arr[17, 30] = 0
-	arr[17, 31] = 0
-	arr[17, 32] = 0
-	arr[17, 33] = 0
-	arr[17, 34] = 0
-	arr[17, 35] = 0
-	arr[17, 36] = 0
-	arr[17, 37] = 0
-	arr[17, 38] = 0
-	arr[17, 39] = 0
-	arr[17, 40] = 0
-	arr[17, 41] = 0
-	arr[17, 42] = 0
-	arr[17, 43] = 0
-	arr[17, 44] = 0
-	arr[17, 45] = 0
-	arr[17, 46] = 0
-	arr[17, 47] = 0
-	arr[17, 48] = -kr[17]
-	arr[17, 49] = kf[17] * y[50]
-	arr[17, 50] = kf[17] * y[49]
-	arr[17, 51] = 0
-	arr[17, 52] = 0
-	arr[17, 53] = 0
-	arr[17, 54] = 0
-	arr[17, 55] = 0
-	arr[17, 56] = 0
-	arr[17, 57] = 0
-	arr[17, 58] = 0
-	arr[17, 59] = 0
-	arr[17, 60] = 0
-	arr[17, 61] = 0
-	arr[17, 62] = 0
-	arr[17, 63] = 0
-	arr[17, 64] = 0
-	arr[17, 65] = 0
-	arr[17, 66] = 0
-	arr[17, 67] = 0
-	arr[17, 68] = 0
-	arr[17, 69] = 0
-	arr[17, 70] = 0
-	arr[17, 71] = 0
-	arr[17, 72] = 0
-	arr[17, 73] = 0
-	arr[17, 74] = 0
-	arr[17, 75] = 0
-	arr[17, 76] = 0
-	arr[17, 77] = 0
-	arr[17, 78] = 0
-	arr[17, 79] = 0
-	arr[17, 80] = 0
-	arr[17, 81] = 0
-	arr[17, 82] = 0
-	arr[17, 83] = 0
-	arr[17, 84] = 0
-	arr[17, 85] = 0
-	arr[17, 86] = 0
-	arr[17, 87] = 0
-	arr[17, 88] = 0
-	arr[17, 89] = 0
-	arr[17, 90] = 0
-	arr[17, 91] = 0
-	arr[17, 92] = 0
-	arr[17, 93] = 0
-	arr[17, 94] = 0
-	arr[17, 95] = 0
-	arr[17, 96] = 0
-	arr[17, 97] = 0
-	arr[17, 98] = 0
-	arr[17, 99] = 0
-	arr[17, 100] = 0
-	arr[17, 101] = 0
-	arr[17, 102] = 0
-	arr[17, 103] = 0
-	arr[18, 0] = 0
-	arr[18, 1] = 0
-	arr[18, 2] = 0
-	arr[18, 3] = 0
-	arr[18, 4] = 0
-	arr[18, 5] = 0
-	arr[18, 6] = 0
-	arr[18, 7] = 0
-	arr[18, 8] = 0
-	arr[18, 9] = 0
-	arr[18, 10] = 0
-	arr[18, 11] = 0
-	arr[18, 12] = 0
-	arr[18, 13] = 0
-	arr[18, 14] = 0
-	arr[18, 15] = 0
-	arr[18, 16] = 0
-	arr[18, 17] = 0
-	arr[18, 18] = 0
-	arr[18, 19] = 0
-	arr[18, 20] = 0
-	arr[18, 21] = 0
-	arr[18, 22] = 0
-	arr[18, 23] = 0
-	arr[18, 24] = 0
-	arr[18, 25] = 0
-	arr[18, 26] = 0
-	arr[18, 27] = 0
-	arr[18, 28] = 0
-	arr[18, 29] = 0
-	arr[18, 30] = 0
-	arr[18, 31] = 0
-	arr[18, 32] = 0
-	arr[18, 33] = 0
-	arr[18, 34] = 0
-	arr[18, 35] = 0
-	arr[18, 36] = 0
-	arr[18, 37] = 0
-	arr[18, 38] = 0
-	arr[18, 39] = 0
-	arr[18, 40] = 0
-	arr[18, 41] = 0
-	arr[18, 42] = 0
-	arr[18, 43] = 0
-	arr[18, 44] = 0
-	arr[18, 45] = 0
-	arr[18, 46] = 0
-	arr[18, 47] = 0
-	arr[18, 48] = 0
-	arr[18, 49] = 0
-	arr[18, 50] = 0
-	arr[18, 51] = -kr[18]
-	arr[18, 52] = kf[18] * y[53]
-	arr[18, 53] = kf[18] * y[52]
-	arr[18, 54] = 0
-	arr[18, 55] = 0
-	arr[18, 56] = 0
-	arr[18, 57] = 0
-	arr[18, 58] = 0
-	arr[18, 59] = 0
-	arr[18, 60] = 0
-	arr[18, 61] = 0
-	arr[18, 62] = 0
-	arr[18, 63] = 0
-	arr[18, 64] = 0
-	arr[18, 65] = 0
-	arr[18, 66] = 0
-	arr[18, 67] = 0
-	arr[18, 68] = 0
-	arr[18, 69] = 0
-	arr[18, 70] = 0
-	arr[18, 71] = 0
-	arr[18, 72] = 0
-	arr[18, 73] = 0
-	arr[18, 74] = 0
-	arr[18, 75] = 0
-	arr[18, 76] = 0
-	arr[18, 77] = 0
-	arr[18, 78] = 0
-	arr[18, 79] = 0
-	arr[18, 80] = 0
-	arr[18, 81] = 0
-	arr[18, 82] = 0
-	arr[18, 83] = 0
-	arr[18, 84] = 0
-	arr[18, 85] = 0
-	arr[18, 86] = 0
-	arr[18, 87] = 0
-	arr[18, 88] = 0
-	arr[18, 89] = 0
-	arr[18, 90] = 0
-	arr[18, 91] = 0
-	arr[18, 92] = 0
-	arr[18, 93] = 0
-	arr[18, 94] = 0
-	arr[18, 95] = 0
-	arr[18, 96] = 0
-	arr[18, 97] = 0
-	arr[18, 98] = 0
-	arr[18, 99] = 0
-	arr[18, 100] = 0
-	arr[18, 101] = 0
-	arr[18, 102] = 0
-	arr[18, 103] = 0
-	arr[19, 0] = 0
-	arr[19, 1] = 0
-	arr[19, 2] = 0
-	arr[19, 3] = 0
-	arr[19, 4] = 0
-	arr[19, 5] = 0
-	arr[19, 6] = 0
-	arr[19, 7] = 0
-	arr[19, 8] = 0
-	arr[19, 9] = 0
-	arr[19, 10] = 0
-	arr[19, 11] = 0
-	arr[19, 12] = 0
-	arr[19, 13] = 0
-	arr[19, 14] = 0
-	arr[19, 15] = 0
-	arr[19, 16] = 0
-	arr[19, 17] = 0
-	arr[19, 18] = 0
-	arr[19, 19] = 0
-	arr[19, 20] = 0
-	arr[19, 21] = 0
-	arr[19, 22] = 0
-	arr[19, 23] = 0
-	arr[19, 24] = 0
-	arr[19, 25] = 0
-	arr[19, 26] = 0
-	arr[19, 27] = 0
-	arr[19, 28] = 0
-	arr[19, 29] = 0
-	arr[19, 30] = 0
-	arr[19, 31] = 0
-	arr[19, 32] = 0
-	arr[19, 33] = 0
-	arr[19, 34] = 0
-	arr[19, 35] = 0
-	arr[19, 36] = 0
-	arr[19, 37] = 0
-	arr[19, 38] = 0
-	arr[19, 39] = 0
-	arr[19, 40] = 0
-	arr[19, 41] = 0
-	arr[19, 42] = 0
-	arr[19, 43] = 0
-	arr[19, 44] = 0
-	arr[19, 45] = 0
-	arr[19, 46] = 0
-	arr[19, 47] = 0
-	arr[19, 48] = 0
-	arr[19, 49] = 0
-	arr[19, 50] = 0
-	arr[19, 51] = 0
-	arr[19, 52] = 0
-	arr[19, 53] = 0
-	arr[19, 54] = -kr[19]
-	arr[19, 55] = kf[19] * y[56]
-	arr[19, 56] = kf[19] * y[55]
-	arr[19, 57] = 0
-	arr[19, 58] = 0
-	arr[19, 59] = 0
-	arr[19, 60] = 0
-	arr[19, 61] = 0
-	arr[19, 62] = 0
-	arr[19, 63] = 0
-	arr[19, 64] = 0
-	arr[19, 65] = 0
-	arr[19, 66] = 0
-	arr[19, 67] = 0
-	arr[19, 68] = 0
-	arr[19, 69] = 0
-	arr[19, 70] = 0
-	arr[19, 71] = 0
-	arr[19, 72] = 0
-	arr[19, 73] = 0
-	arr[19, 74] = 0
-	arr[19, 75] = 0
-	arr[19, 76] = 0
-	arr[19, 77] = 0
-	arr[19, 78] = 0
-	arr[19, 79] = 0
-	arr[19, 80] = 0
-	arr[19, 81] = 0
-	arr[19, 82] = 0
-	arr[19, 83] = 0
-	arr[19, 84] = 0
-	arr[19, 85] = 0
-	arr[19, 86] = 0
-	arr[19, 87] = 0
-	arr[19, 88] = 0
-	arr[19, 89] = 0
-	arr[19, 90] = 0
-	arr[19, 91] = 0
-	arr[19, 92] = 0
-	arr[19, 93] = 0
-	arr[19, 94] = 0
-	arr[19, 95] = 0
-	arr[19, 96] = 0
-	arr[19, 97] = 0
-	arr[19, 98] = 0
-	arr[19, 99] = 0
-	arr[19, 100] = 0
-	arr[19, 101] = 0
-	arr[19, 102] = 0
-	arr[19, 103] = 0
-	arr[20, 0] = 0
-	arr[20, 1] = 0
-	arr[20, 2] = 0
-	arr[20, 3] = 0
-	arr[20, 4] = 0
-	arr[20, 5] = 0
-	arr[20, 6] = 0
-	arr[20, 7] = 0
-	arr[20, 8] = 0
-	arr[20, 9] = 0
-	arr[20, 10] = 0
-	arr[20, 11] = 0
-	arr[20, 12] = 0
-	arr[20, 13] = 0
-	arr[20, 14] = 0
-	arr[20, 15] = 0
-	arr[20, 16] = 0
-	arr[20, 17] = 0
-	arr[20, 18] = 0
-	arr[20, 19] = 0
-	arr[20, 20] = 0
-	arr[20, 21] = 0
-	arr[20, 22] = kf[20] * y[58]
-	arr[20, 23] = 0
-	arr[20, 24] = 0
-	arr[20, 25] = 0
-	arr[20, 26] = 0
-	arr[20, 27] = 0
-	arr[20, 28] = 0
-	arr[20, 29] = 0
-	arr[20, 30] = 0
-	arr[20, 31] = 0
-	arr[20, 32] = 0
-	arr[20, 33] = 0
-	arr[20, 34] = 0
-	arr[20, 35] = 0
-	arr[20, 36] = 0
-	arr[20, 37] = 0
-	arr[20, 38] = 0
-	arr[20, 39] = 0
-	arr[20, 40] = 0
-	arr[20, 41] = 0
-	arr[20, 42] = 0
-	arr[20, 43] = 0
-	arr[20, 44] = 0
-	arr[20, 45] = 0
-	arr[20, 46] = 0
-	arr[20, 47] = 0
-	arr[20, 48] = 0
-	arr[20, 49] = 0
-	arr[20, 50] = 0
-	arr[20, 51] = 0
-	arr[20, 52] = 0
-	arr[20, 53] = 0
-	arr[20, 54] = 0
-	arr[20, 55] = 0
-	arr[20, 56] = 0
-	arr[20, 57] = -kr[20]
-	arr[20, 58] = kf[20] * y[22]
-	arr[20, 59] = 0
-	arr[20, 60] = 0
-	arr[20, 61] = 0
-	arr[20, 62] = 0
-	arr[20, 63] = 0
-	arr[20, 64] = 0
-	arr[20, 65] = 0
-	arr[20, 66] = 0
-	arr[20, 67] = 0
-	arr[20, 68] = 0
-	arr[20, 69] = 0
-	arr[20, 70] = 0
-	arr[20, 71] = 0
-	arr[20, 72] = 0
-	arr[20, 73] = 0
-	arr[20, 74] = 0
-	arr[20, 75] = 0
-	arr[20, 76] = 0
-	arr[20, 77] = 0
-	arr[20, 78] = 0
-	arr[20, 79] = 0
-	arr[20, 80] = 0
-	arr[20, 81] = 0
-	arr[20, 82] = 0
-	arr[20, 83] = 0
-	arr[20, 84] = 0
-	arr[20, 85] = 0
-	arr[20, 86] = 0
-	arr[20, 87] = 0
-	arr[20, 88] = 0
-	arr[20, 89] = 0
-	arr[20, 90] = 0
-	arr[20, 91] = 0
-	arr[20, 92] = 0
-	arr[20, 93] = 0
-	arr[20, 94] = 0
-	arr[20, 95] = 0
-	arr[20, 96] = 0
-	arr[20, 97] = 0
-	arr[20, 98] = 0
-	arr[20, 99] = 0
-	arr[20, 100] = 0
-	arr[20, 101] = 0
-	arr[20, 102] = 0
-	arr[20, 103] = 0
-	arr[21, 0] = 0
-	arr[21, 1] = 0
-	arr[21, 2] = 0
-	arr[21, 3] = 0
-	arr[21, 4] = 0
-	arr[21, 5] = 0
-	arr[21, 6] = 0
-	arr[21, 7] = 0
-	arr[21, 8] = 0
-	arr[21, 9] = 0
-	arr[21, 10] = 0
-	arr[21, 11] = 0
-	arr[21, 12] = 0
-	arr[21, 13] = 0
-	arr[21, 14] = 0
-	arr[21, 15] = 0
-	arr[21, 16] = 0
-	arr[21, 17] = 0
-	arr[21, 18] = 0
-	arr[21, 19] = 0
-	arr[21, 20] = 0
-	arr[21, 21] = 0
-	arr[21, 22] = 0
-	arr[21, 23] = 0
-	arr[21, 24] = 0
-	arr[21, 25] = 0
-	arr[21, 26] = 0
-	arr[21, 27] = 0
-	arr[21, 28] = 0
-	arr[21, 29] = 0
-	arr[21, 30] = 0
-	arr[21, 31] = 0
-	arr[21, 32] = 0
-	arr[21, 33] = 0
-	arr[21, 34] = 0
-	arr[21, 35] = 0
-	arr[21, 36] = 0
-	arr[21, 37] = 0
-	arr[21, 38] = 0
-	arr[21, 39] = 0
-	arr[21, 40] = 0
-	arr[21, 41] = 0
-	arr[21, 42] = 0
-	arr[21, 43] = 0
-	arr[21, 44] = 0
-	arr[21, 45] = 0
-	arr[21, 46] = 0
-	arr[21, 47] = 0
-	arr[21, 48] = 0
-	arr[21, 49] = 0
-	arr[21, 50] = 0
-	arr[21, 51] = 0
-	arr[21, 52] = 0
-	arr[21, 53] = 0
-	arr[21, 54] = 0
-	arr[21, 55] = 0
-	arr[21, 56] = 0
-	arr[21, 57] = 0
-	arr[21, 58] = 0
-	arr[21, 59] = -kr[21]
-	arr[21, 60] = kf[21] * y[61]
-	arr[21, 61] = kf[21] * y[60]
-	arr[21, 62] = 0
-	arr[21, 63] = 0
-	arr[21, 64] = 0
-	arr[21, 65] = 0
-	arr[21, 66] = 0
-	arr[21, 67] = 0
-	arr[21, 68] = 0
-	arr[21, 69] = 0
-	arr[21, 70] = 0
-	arr[21, 71] = 0
-	arr[21, 72] = 0
-	arr[21, 73] = 0
-	arr[21, 74] = 0
-	arr[21, 75] = 0
-	arr[21, 76] = 0
-	arr[21, 77] = 0
-	arr[21, 78] = 0
-	arr[21, 79] = 0
-	arr[21, 80] = 0
-	arr[21, 81] = 0
-	arr[21, 82] = 0
-	arr[21, 83] = 0
-	arr[21, 84] = 0
-	arr[21, 85] = 0
-	arr[21, 86] = 0
-	arr[21, 87] = 0
-	arr[21, 88] = 0
-	arr[21, 89] = 0
-	arr[21, 90] = 0
-	arr[21, 91] = 0
-	arr[21, 92] = 0
-	arr[21, 93] = 0
-	arr[21, 94] = 0
-	arr[21, 95] = 0
-	arr[21, 96] = 0
-	arr[21, 97] = 0
-	arr[21, 98] = 0
-	arr[21, 99] = 0
-	arr[21, 100] = 0
-	arr[21, 101] = 0
-	arr[21, 102] = 0
-	arr[21, 103] = 0
-	arr[22, 0] = 0
-	arr[22, 1] = 0
-	arr[22, 2] = 0
-	arr[22, 3] = 0
-	arr[22, 4] = 0
-	arr[22, 5] = 0
-	arr[22, 6] = 0
-	arr[22, 7] = 0
-	arr[22, 8] = 0
-	arr[22, 9] = 0
-	arr[22, 10] = 0
-	arr[22, 11] = 0
-	arr[22, 12] = 0
-	arr[22, 13] = 0
-	arr[22, 14] = 0
-	arr[22, 15] = 0
-	arr[22, 16] = 0
-	arr[22, 17] = 0
-	arr[22, 18] = 0
-	arr[22, 19] = 0
-	arr[22, 20] = 0
-	arr[22, 21] = 0
-	arr[22, 22] = 0
-	arr[22, 23] = 0
-	arr[22, 24] = 0
-	arr[22, 25] = 0
-	arr[22, 26] = 0
-	arr[22, 27] = 0
-	arr[22, 28] = 0
-	arr[22, 29] = 0
-	arr[22, 30] = 0
-	arr[22, 31] = 0
-	arr[22, 32] = 0
-	arr[22, 33] = 0
-	arr[22, 34] = 0
-	arr[22, 35] = 0
-	arr[22, 36] = 0
-	arr[22, 37] = 0
-	arr[22, 38] = 0
-	arr[22, 39] = 0
-	arr[22, 40] = 0
-	arr[22, 41] = 0
-	arr[22, 42] = 0
-	arr[22, 43] = 0
-	arr[22, 44] = 0
-	arr[22, 45] = 0
-	arr[22, 46] = 0
-	arr[22, 47] = 0
-	arr[22, 48] = 0
-	arr[22, 49] = 0
-	arr[22, 50] = 0
-	arr[22, 51] = 0
-	arr[22, 52] = 0
-	arr[22, 53] = 0
-	arr[22, 54] = 0
-	arr[22, 55] = 0
-	arr[22, 56] = 0
-	arr[22, 57] = 0
-	arr[22, 58] = 0
-	arr[22, 59] = 0
-	arr[22, 60] = 0
-	arr[22, 61] = 0
-	arr[22, 62] = -kr[22]
-	arr[22, 63] = kf[22] * y[64]
-	arr[22, 64] = kf[22] * y[63]
-	arr[22, 65] = 0
-	arr[22, 66] = 0
-	arr[22, 67] = 0
-	arr[22, 68] = 0
-	arr[22, 69] = 0
-	arr[22, 70] = 0
-	arr[22, 71] = 0
-	arr[22, 72] = 0
-	arr[22, 73] = 0
-	arr[22, 74] = 0
-	arr[22, 75] = 0
-	arr[22, 76] = 0
-	arr[22, 77] = 0
-	arr[22, 78] = 0
-	arr[22, 79] = 0
-	arr[22, 80] = 0
-	arr[22, 81] = 0
-	arr[22, 82] = 0
-	arr[22, 83] = 0
-	arr[22, 84] = 0
-	arr[22, 85] = 0
-	arr[22, 86] = 0
-	arr[22, 87] = 0
-	arr[22, 88] = 0
-	arr[22, 89] = 0
-	arr[22, 90] = 0
-	arr[22, 91] = 0
-	arr[22, 92] = 0
-	arr[22, 93] = 0
-	arr[22, 94] = 0
-	arr[22, 95] = 0
-	arr[22, 96] = 0
-	arr[22, 97] = 0
-	arr[22, 98] = 0
-	arr[22, 99] = 0
-	arr[22, 100] = 0
-	arr[22, 101] = 0
-	arr[22, 102] = 0
-	arr[22, 103] = 0
-	arr[23, 0] = 0
-	arr[23, 1] = 0
-	arr[23, 2] = 0
-	arr[23, 3] = 0
-	arr[23, 4] = 0
-	arr[23, 5] = 0
-	arr[23, 6] = 0
-	arr[23, 7] = 0
-	arr[23, 8] = 0
-	arr[23, 9] = 0
-	arr[23, 10] = 0
-	arr[23, 11] = 0
-	arr[23, 12] = 0
-	arr[23, 13] = 0
-	arr[23, 14] = 0
-	arr[23, 15] = 0
-	arr[23, 16] = 0
-	arr[23, 17] = 0
-	arr[23, 18] = 0
-	arr[23, 19] = 0
-	arr[23, 20] = 0
-	arr[23, 21] = 0
-	arr[23, 22] = 0
-	arr[23, 23] = 0
-	arr[23, 24] = 0
-	arr[23, 25] = 0
-	arr[23, 26] = 0
-	arr[23, 27] = 0
-	arr[23, 28] = 0
-	arr[23, 29] = 0
-	arr[23, 30] = 0
-	arr[23, 31] = 0
-	arr[23, 32] = 0
-	arr[23, 33] = 0
-	arr[23, 34] = 0
-	arr[23, 35] = 0
-	arr[23, 36] = 0
-	arr[23, 37] = 0
-	arr[23, 38] = 0
-	arr[23, 39] = 0
-	arr[23, 40] = 0
-	arr[23, 41] = 0
-	arr[23, 42] = 0
-	arr[23, 43] = 0
-	arr[23, 44] = 0
-	arr[23, 45] = 0
-	arr[23, 46] = 0
-	arr[23, 47] = 0
-	arr[23, 48] = 0
-	arr[23, 49] = 0
-	arr[23, 50] = 0
-	arr[23, 51] = 0
-	arr[23, 52] = 0
-	arr[23, 53] = 0
-	arr[23, 54] = 0
-	arr[23, 55] = 0
-	arr[23, 56] = 0
-	arr[23, 57] = 0
-	arr[23, 58] = 0
-	arr[23, 59] = 0
-	arr[23, 60] = 0
-	arr[23, 61] = 0
-	arr[23, 62] = 0
-	arr[23, 63] = 0
-	arr[23, 64] = 0
-	arr[23, 65] = -kr[23]
-	arr[23, 66] = kf[23] * y[67]
-	arr[23, 67] = kf[23] * y[66]
-	arr[23, 68] = 0
-	arr[23, 69] = 0
-	arr[23, 70] = 0
-	arr[23, 71] = 0
-	arr[23, 72] = 0
-	arr[23, 73] = 0
-	arr[23, 74] = 0
-	arr[23, 75] = 0
-	arr[23, 76] = 0
-	arr[23, 77] = 0
-	arr[23, 78] = 0
-	arr[23, 79] = 0
-	arr[23, 80] = 0
-	arr[23, 81] = 0
-	arr[23, 82] = 0
-	arr[23, 83] = 0
-	arr[23, 84] = 0
-	arr[23, 85] = 0
-	arr[23, 86] = 0
-	arr[23, 87] = 0
-	arr[23, 88] = 0
-	arr[23, 89] = 0
-	arr[23, 90] = 0
-	arr[23, 91] = 0
-	arr[23, 92] = 0
-	arr[23, 93] = 0
-	arr[23, 94] = 0
-	arr[23, 95] = 0
-	arr[23, 96] = 0
-	arr[23, 97] = 0
-	arr[23, 98] = 0
-	arr[23, 99] = 0
-	arr[23, 100] = 0
-	arr[23, 101] = 0
-	arr[23, 102] = 0
-	arr[23, 103] = 0
-	arr[24, 0] = 0
-	arr[24, 1] = 0
-	arr[24, 2] = 0
-	arr[24, 3] = 0
-	arr[24, 4] = 0
-	arr[24, 5] = 0
-	arr[24, 6] = 0
-	arr[24, 7] = 0
-	arr[24, 8] = 0
-	arr[24, 9] = 0
-	arr[24, 10] = 0
-	arr[24, 11] = 0
-	arr[24, 12] = 0
-	arr[24, 13] = 0
-	arr[24, 14] = 0
-	arr[24, 15] = 0
-	arr[24, 16] = 0
-	arr[24, 17] = 0
-	arr[24, 18] = 0
-	arr[24, 19] = 0
-	arr[24, 20] = 0
-	arr[24, 21] = 0
-	arr[24, 22] = 0
-	arr[24, 23] = 0
-	arr[24, 24] = 0
-	arr[24, 25] = 0
-	arr[24, 26] = 0
-	arr[24, 27] = 0
-	arr[24, 28] = 0
-	arr[24, 29] = 0
-	arr[24, 30] = 0
-	arr[24, 31] = 0
-	arr[24, 32] = 0
-	arr[24, 33] = 0
-	arr[24, 34] = 0
-	arr[24, 35] = 0
-	arr[24, 36] = 0
-	arr[24, 37] = 0
-	arr[24, 38] = 0
-	arr[24, 39] = 0
-	arr[24, 40] = 0
-	arr[24, 41] = 0
-	arr[24, 42] = 0
-	arr[24, 43] = 0
-	arr[24, 44] = 0
-	arr[24, 45] = 0
-	arr[24, 46] = 0
-	arr[24, 47] = 0
-	arr[24, 48] = 0
-	arr[24, 49] = 0
-	arr[24, 50] = 0
-	arr[24, 51] = 0
-	arr[24, 52] = 0
-	arr[24, 53] = 0
-	arr[24, 54] = 0
-	arr[24, 55] = 0
-	arr[24, 56] = 0
-	arr[24, 57] = 0
-	arr[24, 58] = 0
-	arr[24, 59] = 0
-	arr[24, 60] = 0
-	arr[24, 61] = 0
-	arr[24, 62] = 0
-	arr[24, 63] = 0
-	arr[24, 64] = 0
-	arr[24, 65] = 0
-	arr[24, 66] = kf[24] * y[69]
-	arr[24, 67] = 0
-	arr[24, 68] = -kr[24]
-	arr[24, 69] = kf[24] * y[66]
-	arr[24, 70] = 0
-	arr[24, 71] = 0
-	arr[24, 72] = 0
-	arr[24, 73] = 0
-	arr[24, 74] = 0
-	arr[24, 75] = 0
-	arr[24, 76] = 0
-	arr[24, 77] = 0
-	arr[24, 78] = 0
-	arr[24, 79] = 0
-	arr[24, 80] = 0
-	arr[24, 81] = 0
-	arr[24, 82] = 0
-	arr[24, 83] = 0
-	arr[24, 84] = 0
-	arr[24, 85] = 0
-	arr[24, 86] = 0
-	arr[24, 87] = 0
-	arr[24, 88] = 0
-	arr[24, 89] = 0
-	arr[24, 90] = 0
-	arr[24, 91] = 0
-	arr[24, 92] = 0
-	arr[24, 93] = 0
-	arr[24, 94] = 0
-	arr[24, 95] = 0
-	arr[24, 96] = 0
-	arr[24, 97] = 0
-	arr[24, 98] = 0
-	arr[24, 99] = 0
-	arr[24, 100] = 0
-	arr[24, 101] = 0
-	arr[24, 102] = 0
-	arr[24, 103] = 0
-	arr[25, 0] = 0
-	arr[25, 1] = 0
-	arr[25, 2] = 0
-	arr[25, 3] = 0
-	arr[25, 4] = 0
-	arr[25, 5] = 0
-	arr[25, 6] = 0
-	arr[25, 7] = 0
-	arr[25, 8] = 0
-	arr[25, 9] = 0
-	arr[25, 10] = 0
-	arr[25, 11] = 0
-	arr[25, 12] = 0
-	arr[25, 13] = 0
-	arr[25, 14] = 0
-	arr[25, 15] = 0
-	arr[25, 16] = 0
-	arr[25, 17] = 0
-	arr[25, 18] = 0
-	arr[25, 19] = 0
-	arr[25, 20] = 0
-	arr[25, 21] = 0
-	arr[25, 22] = 0
-	arr[25, 23] = 0
-	arr[25, 24] = 0
-	arr[25, 25] = 0
-	arr[25, 26] = 0
-	arr[25, 27] = 0
-	arr[25, 28] = 0
-	arr[25, 29] = 0
-	arr[25, 30] = 0
-	arr[25, 31] = 0
-	arr[25, 32] = 0
-	arr[25, 33] = 0
-	arr[25, 34] = 0
-	arr[25, 35] = 0
-	arr[25, 36] = 0
-	arr[25, 37] = 0
-	arr[25, 38] = 0
-	arr[25, 39] = 0
-	arr[25, 40] = 0
-	arr[25, 41] = 0
-	arr[25, 42] = 0
-	arr[25, 43] = 0
-	arr[25, 44] = 0
-	arr[25, 45] = 0
-	arr[25, 46] = 0
-	arr[25, 47] = 0
-	arr[25, 48] = 0
-	arr[25, 49] = 0
-	arr[25, 50] = 0
-	arr[25, 51] = 0
-	arr[25, 52] = 0
-	arr[25, 53] = 0
-	arr[25, 54] = 0
-	arr[25, 55] = 0
-	arr[25, 56] = 0
-	arr[25, 57] = 0
-	arr[25, 58] = 0
-	arr[25, 59] = 0
-	arr[25, 60] = 0
-	arr[25, 61] = 0
-	arr[25, 62] = 0
-	arr[25, 63] = kf[25] * y[71]
-	arr[25, 64] = 0
-	arr[25, 65] = 0
-	arr[25, 66] = 0
-	arr[25, 67] = 0
-	arr[25, 68] = 0
-	arr[25, 69] = 0
-	arr[25, 70] = -kr[25]
-	arr[25, 71] = kf[25] * y[63]
-	arr[25, 72] = 0
-	arr[25, 73] = 0
-	arr[25, 74] = 0
-	arr[25, 75] = 0
-	arr[25, 76] = 0
-	arr[25, 77] = 0
-	arr[25, 78] = 0
-	arr[25, 79] = 0
-	arr[25, 80] = 0
-	arr[25, 81] = 0
-	arr[25, 82] = 0
-	arr[25, 83] = 0
-	arr[25, 84] = 0
-	arr[25, 85] = 0
-	arr[25, 86] = 0
-	arr[25, 87] = 0
-	arr[25, 88] = 0
-	arr[25, 89] = 0
-	arr[25, 90] = 0
-	arr[25, 91] = 0
-	arr[25, 92] = 0
-	arr[25, 93] = 0
-	arr[25, 94] = 0
-	arr[25, 95] = 0
-	arr[25, 96] = 0
-	arr[25, 97] = 0
-	arr[25, 98] = 0
-	arr[25, 99] = 0
-	arr[25, 100] = 0
-	arr[25, 101] = 0
-	arr[25, 102] = 0
-	arr[25, 103] = 0
-	arr[26, 0] = 0
-	arr[26, 1] = 0
-	arr[26, 2] = 0
-	arr[26, 3] = 0
-	arr[26, 4] = 0
-	arr[26, 5] = 0
-	arr[26, 6] = 0
-	arr[26, 7] = 0
-	arr[26, 8] = 0
-	arr[26, 9] = 0
-	arr[26, 10] = 0
-	arr[26, 11] = 0
-	arr[26, 12] = 0
-	arr[26, 13] = 0
-	arr[26, 14] = 0
-	arr[26, 15] = 0
-	arr[26, 16] = 0
-	arr[26, 17] = 0
-	arr[26, 18] = 0
-	arr[26, 19] = 0
-	arr[26, 20] = 0
-	arr[26, 21] = 0
-	arr[26, 22] = 0
-	arr[26, 23] = 0
-	arr[26, 24] = 0
-	arr[26, 25] = 0
-	arr[26, 26] = 0
-	arr[26, 27] = 0
-	arr[26, 28] = 0
-	arr[26, 29] = 0
-	arr[26, 30] = 0
-	arr[26, 31] = 0
-	arr[26, 32] = 0
-	arr[26, 33] = 0
-	arr[26, 34] = 0
-	arr[26, 35] = 0
-	arr[26, 36] = 0
-	arr[26, 37] = 0
-	arr[26, 38] = 0
-	arr[26, 39] = 0
-	arr[26, 40] = kf[26] * y[73]
-	arr[26, 41] = 0
-	arr[26, 42] = 0
-	arr[26, 43] = 0
-	arr[26, 44] = 0
-	arr[26, 45] = 0
-	arr[26, 46] = 0
-	arr[26, 47] = 0
-	arr[26, 48] = 0
-	arr[26, 49] = 0
-	arr[26, 50] = 0
-	arr[26, 51] = 0
-	arr[26, 52] = 0
-	arr[26, 53] = 0
-	arr[26, 54] = 0
-	arr[26, 55] = 0
-	arr[26, 56] = 0
-	arr[26, 57] = 0
-	arr[26, 58] = 0
-	arr[26, 59] = 0
-	arr[26, 60] = 0
-	arr[26, 61] = 0
-	arr[26, 62] = 0
-	arr[26, 63] = 0
-	arr[26, 64] = 0
-	arr[26, 65] = 0
-	arr[26, 66] = 0
-	arr[26, 67] = 0
-	arr[26, 68] = 0
-	arr[26, 69] = 0
-	arr[26, 70] = 0
-	arr[26, 71] = 0
-	arr[26, 72] = -kr[26]
-	arr[26, 73] = kf[26] * y[40]
-	arr[26, 74] = 0
-	arr[26, 75] = 0
-	arr[26, 76] = 0
-	arr[26, 77] = 0
-	arr[26, 78] = 0
-	arr[26, 79] = 0
-	arr[26, 80] = 0
-	arr[26, 81] = 0
-	arr[26, 82] = 0
-	arr[26, 83] = 0
-	arr[26, 84] = 0
-	arr[26, 85] = 0
-	arr[26, 86] = 0
-	arr[26, 87] = 0
-	arr[26, 88] = 0
-	arr[26, 89] = 0
-	arr[26, 90] = 0
-	arr[26, 91] = 0
-	arr[26, 92] = 0
-	arr[26, 93] = 0
-	arr[26, 94] = 0
-	arr[26, 95] = 0
-	arr[26, 96] = 0
-	arr[26, 97] = 0
-	arr[26, 98] = 0
-	arr[26, 99] = 0
-	arr[26, 100] = 0
-	arr[26, 101] = 0
-	arr[26, 102] = 0
-	arr[26, 103] = 0
-	arr[27, 0] = 0
-	arr[27, 1] = 0
-	arr[27, 2] = 0
-	arr[27, 3] = 0
-	arr[27, 4] = 0
-	arr[27, 5] = 0
-	arr[27, 6] = 0
-	arr[27, 7] = 0
-	arr[27, 8] = 0
-	arr[27, 9] = 0
-	arr[27, 10] = 0
-	arr[27, 11] = 0
-	arr[27, 12] = 0
-	arr[27, 13] = 0
-	arr[27, 14] = 0
-	arr[27, 15] = 0
-	arr[27, 16] = 0
-	arr[27, 17] = 0
-	arr[27, 18] = 0
-	arr[27, 19] = 0
-	arr[27, 20] = 0
-	arr[27, 21] = 0
-	arr[27, 22] = 0
-	arr[27, 23] = 0
-	arr[27, 24] = 0
-	arr[27, 25] = 0
-	arr[27, 26] = 0
-	arr[27, 27] = 0
-	arr[27, 28] = 0
-	arr[27, 29] = 0
-	arr[27, 30] = 0
-	arr[27, 31] = 0
-	arr[27, 32] = 0
-	arr[27, 33] = 0
-	arr[27, 34] = 0
-	arr[27, 35] = 0
-	arr[27, 36] = 0
-	arr[27, 37] = 0
-	arr[27, 38] = 0
-	arr[27, 39] = 0
-	arr[27, 40] = 0
-	arr[27, 41] = 0
-	arr[27, 42] = 0
-	arr[27, 43] = 0
-	arr[27, 44] = 0
-	arr[27, 45] = 0
-	arr[27, 46] = 0
-	arr[27, 47] = 0
-	arr[27, 48] = 0
-	arr[27, 49] = 0
-	arr[27, 50] = 0
-	arr[27, 51] = 0
-	arr[27, 52] = 0
-	arr[27, 53] = 0
-	arr[27, 54] = 0
-	arr[27, 55] = 0
-	arr[27, 56] = 0
-	arr[27, 57] = 0
-	arr[27, 58] = 0
-	arr[27, 59] = 0
-	arr[27, 60] = 0
-	arr[27, 61] = 0
-	arr[27, 62] = 0
-	arr[27, 63] = 0
-	arr[27, 64] = 0
-	arr[27, 65] = 0
-	arr[27, 66] = 0
-	arr[27, 67] = 0
-	arr[27, 68] = 0
-	arr[27, 69] = 0
-	arr[27, 70] = 0
-	arr[27, 71] = 0
-	arr[27, 72] = 0
-	arr[27, 73] = 0
-	arr[27, 74] = -kr[27] ** 2.0
-	arr[27, 75] = kf[27] * y[76] ** 2.0
-	arr[27, 76] = 2.0 * kf[27] * y[75] * y[76] ** 1.0
-	arr[27, 77] = 0
-	arr[27, 78] = 0
-	arr[27, 79] = 0
-	arr[27, 80] = 0
-	arr[27, 81] = 0
-	arr[27, 82] = 0
-	arr[27, 83] = 0
-	arr[27, 84] = 0
-	arr[27, 85] = 0
-	arr[27, 86] = 0
-	arr[27, 87] = 0
-	arr[27, 88] = 0
-	arr[27, 89] = 0
-	arr[27, 90] = 0
-	arr[27, 91] = 0
-	arr[27, 92] = 0
-	arr[27, 93] = 0
-	arr[27, 94] = 0
-	arr[27, 95] = 0
-	arr[27, 96] = 0
-	arr[27, 97] = 0
-	arr[27, 98] = 0
-	arr[27, 99] = 0
-	arr[27, 100] = 0
-	arr[27, 101] = 0
-	arr[27, 102] = 0
-	arr[27, 103] = 0
-	arr[28, 0] = 0
-	arr[28, 1] = 0
-	arr[28, 2] = 0
-	arr[28, 3] = 0
-	arr[28, 4] = 0
-	arr[28, 5] = 0
-	arr[28, 6] = 0
-	arr[28, 7] = 0
-	arr[28, 8] = 0
-	arr[28, 9] = 0
-	arr[28, 10] = 0
-	arr[28, 11] = 0
-	arr[28, 12] = 0
-	arr[28, 13] = 0
-	arr[28, 14] = 0
-	arr[28, 15] = 0
-	arr[28, 16] = 0
-	arr[28, 17] = 0
-	arr[28, 18] = 0
-	arr[28, 19] = 0
-	arr[28, 20] = 0
-	arr[28, 21] = 0
-	arr[28, 22] = 0
-	arr[28, 23] = 0
-	arr[28, 24] = 0
-	arr[28, 25] = 0
-	arr[28, 26] = 0
-	arr[28, 27] = 0
-	arr[28, 28] = 0
-	arr[28, 29] = 0
-	arr[28, 30] = 0
-	arr[28, 31] = 0
-	arr[28, 32] = 0
-	arr[28, 33] = 0
-	arr[28, 34] = 0
-	arr[28, 35] = 0
-	arr[28, 36] = 0
-	arr[28, 37] = 0
-	arr[28, 38] = 0
-	arr[28, 39] = 0
-	arr[28, 40] = 0
-	arr[28, 41] = 0
-	arr[28, 42] = 0
-	arr[28, 43] = 0
-	arr[28, 44] = 0
-	arr[28, 45] = 0
-	arr[28, 46] = 0
-	arr[28, 47] = 0
-	arr[28, 48] = 0
-	arr[28, 49] = 0
-	arr[28, 50] = 0
-	arr[28, 51] = 0
-	arr[28, 52] = 0
-	arr[28, 53] = 0
-	arr[28, 54] = 0
-	arr[28, 55] = 0
-	arr[28, 56] = 0
-	arr[28, 57] = 0
-	arr[28, 58] = 0
-	arr[28, 59] = 0
-	arr[28, 60] = 0
-	arr[28, 61] = 0
-	arr[28, 62] = 0
-	arr[28, 63] = 0
-	arr[28, 64] = 0
-	arr[28, 65] = 0
-	arr[28, 66] = 0
-	arr[28, 67] = 0
-	arr[28, 68] = 0
-	arr[28, 69] = 0
-	arr[28, 70] = 0
-	arr[28, 71] = 0
-	arr[28, 72] = 0
-	arr[28, 73] = 0
-	arr[28, 74] = 0
-	arr[28, 75] = 0
-	arr[28, 76] = 0
-	arr[28, 77] = -kr[28]
-	arr[28, 78] = kf[28] * y[79]
-	arr[28, 79] = kf[28] * y[78]
-	arr[28, 80] = 0
-	arr[28, 81] = 0
-	arr[28, 82] = 0
-	arr[28, 83] = 0
-	arr[28, 84] = 0
-	arr[28, 85] = 0
-	arr[28, 86] = 0
-	arr[28, 87] = 0
-	arr[28, 88] = 0
-	arr[28, 89] = 0
-	arr[28, 90] = 0
-	arr[28, 91] = 0
-	arr[28, 92] = 0
-	arr[28, 93] = 0
-	arr[28, 94] = 0
-	arr[28, 95] = 0
-	arr[28, 96] = 0
-	arr[28, 97] = 0
-	arr[28, 98] = 0
-	arr[28, 99] = 0
-	arr[28, 100] = 0
-	arr[28, 101] = 0
-	arr[28, 102] = 0
-	arr[28, 103] = 0
-	arr[29, 0] = 0
-	arr[29, 1] = 0
-	arr[29, 2] = 0
-	arr[29, 3] = 0
-	arr[29, 4] = 0
-	arr[29, 5] = 0
-	arr[29, 6] = 0
-	arr[29, 7] = 0
-	arr[29, 8] = 0
-	arr[29, 9] = 0
-	arr[29, 10] = 0
-	arr[29, 11] = 0
-	arr[29, 12] = 0
-	arr[29, 13] = 0
-	arr[29, 14] = 0
-	arr[29, 15] = 0
-	arr[29, 16] = 0
-	arr[29, 17] = 0
-	arr[29, 18] = 0
-	arr[29, 19] = 0
-	arr[29, 20] = 0
-	arr[29, 21] = 0
-	arr[29, 22] = 0
-	arr[29, 23] = 0
-	arr[29, 24] = 0
-	arr[29, 25] = 0
-	arr[29, 26] = 0
-	arr[29, 27] = 0
-	arr[29, 28] = 0
-	arr[29, 29] = 0
-	arr[29, 30] = 0
-	arr[29, 31] = 0
-	arr[29, 32] = 0
-	arr[29, 33] = 0
-	arr[29, 34] = 0
-	arr[29, 35] = 0
-	arr[29, 36] = 0
-	arr[29, 37] = 0
-	arr[29, 38] = 0
-	arr[29, 39] = 0
-	arr[29, 40] = 0
-	arr[29, 41] = 0
-	arr[29, 42] = 0
-	arr[29, 43] = 0
-	arr[29, 44] = 0
-	arr[29, 45] = 0
-	arr[29, 46] = 0
-	arr[29, 47] = 0
-	arr[29, 48] = 0
-	arr[29, 49] = 0
-	arr[29, 50] = 0
-	arr[29, 51] = 0
-	arr[29, 52] = 0
-	arr[29, 53] = 0
-	arr[29, 54] = 0
-	arr[29, 55] = 0
-	arr[29, 56] = 0
-	arr[29, 57] = 0
-	arr[29, 58] = 0
-	arr[29, 59] = 0
-	arr[29, 60] = 0
-	arr[29, 61] = 0
-	arr[29, 62] = 0
-	arr[29, 63] = 0
-	arr[29, 64] = 0
-	arr[29, 65] = 0
-	arr[29, 66] = 0
-	arr[29, 67] = 0
-	arr[29, 68] = 0
-	arr[29, 69] = 0
-	arr[29, 70] = 0
-	arr[29, 71] = 0
-	arr[29, 72] = 0
-	arr[29, 73] = 0
-	arr[29, 74] = 0
-	arr[29, 75] = 0
-	arr[29, 76] = 0
-	arr[29, 77] = 0
-	arr[29, 78] = 0
-	arr[29, 79] = 0
-	arr[29, 80] = -kr[29] ** 2.0
-	arr[29, 81] = kf[29] * y[82] ** 2.0
-	arr[29, 82] = 2.0 * kf[29] * y[81] * y[82] ** 1.0
-	arr[29, 83] = 0
-	arr[29, 84] = 0
-	arr[29, 85] = 0
-	arr[29, 86] = 0
-	arr[29, 87] = 0
-	arr[29, 88] = 0
-	arr[29, 89] = 0
-	arr[29, 90] = 0
-	arr[29, 91] = 0
-	arr[29, 92] = 0
-	arr[29, 93] = 0
-	arr[29, 94] = 0
-	arr[29, 95] = 0
-	arr[29, 96] = 0
-	arr[29, 97] = 0
-	arr[29, 98] = 0
-	arr[29, 99] = 0
-	arr[29, 100] = 0
-	arr[29, 101] = 0
-	arr[29, 102] = 0
-	arr[29, 103] = 0
-	arr[30, 0] = 0
-	arr[30, 1] = 0
-	arr[30, 2] = 0
-	arr[30, 3] = 0
-	arr[30, 4] = 0
-	arr[30, 5] = 0
-	arr[30, 6] = 0
-	arr[30, 7] = 0
-	arr[30, 8] = 0
-	arr[30, 9] = 0
-	arr[30, 10] = 0
-	arr[30, 11] = 0
-	arr[30, 12] = 0
-	arr[30, 13] = 0
-	arr[30, 14] = 0
-	arr[30, 15] = 0
-	arr[30, 16] = 0
-	arr[30, 17] = 0
-	arr[30, 18] = 0
-	arr[30, 19] = 0
-	arr[30, 20] = 0
-	arr[30, 21] = 0
-	arr[30, 22] = 0
-	arr[30, 23] = 0
-	arr[30, 24] = 0
-	arr[30, 25] = 0
-	arr[30, 26] = 0
-	arr[30, 27] = 0
-	arr[30, 28] = 0
-	arr[30, 29] = 0
-	arr[30, 30] = 0
-	arr[30, 31] = 0
-	arr[30, 32] = 0
-	arr[30, 33] = 0
-	arr[30, 34] = 0
-	arr[30, 35] = 0
-	arr[30, 36] = 0
-	arr[30, 37] = 0
-	arr[30, 38] = 0
-	arr[30, 39] = 0
-	arr[30, 40] = 0
-	arr[30, 41] = 0
-	arr[30, 42] = 0
-	arr[30, 43] = 0
-	arr[30, 44] = 0
-	arr[30, 45] = 0
-	arr[30, 46] = 0
-	arr[30, 47] = 0
-	arr[30, 48] = 0
-	arr[30, 49] = 0
-	arr[30, 50] = 0
-	arr[30, 51] = 0
-	arr[30, 52] = 0
-	arr[30, 53] = 0
-	arr[30, 54] = 0
-	arr[30, 55] = 0
-	arr[30, 56] = 0
-	arr[30, 57] = 0
-	arr[30, 58] = 0
-	arr[30, 59] = 0
-	arr[30, 60] = 0
-	arr[30, 61] = 0
-	arr[30, 62] = 0
-	arr[30, 63] = 0
-	arr[30, 64] = 0
-	arr[30, 65] = 0
-	arr[30, 66] = 0
-	arr[30, 67] = 0
-	arr[30, 68] = 0
-	arr[30, 69] = 0
-	arr[30, 70] = 0
-	arr[30, 71] = 0
-	arr[30, 72] = 0
-	arr[30, 73] = 0
-	arr[30, 74] = 0
-	arr[30, 75] = 0
-	arr[30, 76] = 0
-	arr[30, 77] = 0
-	arr[30, 78] = 0
-	arr[30, 79] = 0
-	arr[30, 80] = 0
-	arr[30, 81] = 0
-	arr[30, 82] = 0
-	arr[30, 83] = -kr[30]
-	arr[30, 84] = kf[30] * y[85]
-	arr[30, 85] = kf[30] * y[84]
-	arr[30, 86] = 0
-	arr[30, 87] = 0
-	arr[30, 88] = 0
-	arr[30, 89] = 0
-	arr[30, 90] = 0
-	arr[30, 91] = 0
-	arr[30, 92] = 0
-	arr[30, 93] = 0
-	arr[30, 94] = 0
-	arr[30, 95] = 0
-	arr[30, 96] = 0
-	arr[30, 97] = 0
-	arr[30, 98] = 0
-	arr[30, 99] = 0
-	arr[30, 100] = 0
-	arr[30, 101] = 0
-	arr[30, 102] = 0
-	arr[30, 103] = 0
-	arr[31, 0] = 0
-	arr[31, 1] = 0
-	arr[31, 2] = 0
-	arr[31, 3] = 0
-	arr[31, 4] = 0
-	arr[31, 5] = 0
-	arr[31, 6] = 0
-	arr[31, 7] = 0
-	arr[31, 8] = 0
-	arr[31, 9] = 0
-	arr[31, 10] = 0
-	arr[31, 11] = 0
-	arr[31, 12] = 0
-	arr[31, 13] = 0
-	arr[31, 14] = 0
-	arr[31, 15] = 0
-	arr[31, 16] = 0
-	arr[31, 17] = 0
-	arr[31, 18] = 0
-	arr[31, 19] = 0
-	arr[31, 20] = 0
-	arr[31, 21] = 0
-	arr[31, 22] = 0
-	arr[31, 23] = 0
-	arr[31, 24] = 0
-	arr[31, 25] = 0
-	arr[31, 26] = 0
-	arr[31, 27] = 0
-	arr[31, 28] = 0
-	arr[31, 29] = 0
-	arr[31, 30] = 0
-	arr[31, 31] = 0
-	arr[31, 32] = 0
-	arr[31, 33] = 0
-	arr[31, 34] = 0
-	arr[31, 35] = 0
-	arr[31, 36] = 0
-	arr[31, 37] = 0
-	arr[31, 38] = 0
-	arr[31, 39] = 0
-	arr[31, 40] = 0
-	arr[31, 41] = 0
-	arr[31, 42] = 0
-	arr[31, 43] = 0
-	arr[31, 44] = 0
-	arr[31, 45] = 0
-	arr[31, 46] = 0
-	arr[31, 47] = 0
-	arr[31, 48] = 0
-	arr[31, 49] = 0
-	arr[31, 50] = 0
-	arr[31, 51] = 0
-	arr[31, 52] = 0
-	arr[31, 53] = 0
-	arr[31, 54] = 0
-	arr[31, 55] = 0
-	arr[31, 56] = 0
-	arr[31, 57] = 0
-	arr[31, 58] = 0
-	arr[31, 59] = 0
-	arr[31, 60] = 0
-	arr[31, 61] = 0
-	arr[31, 62] = 0
-	arr[31, 63] = 0
-	arr[31, 64] = 0
-	arr[31, 65] = 0
-	arr[31, 66] = 0
-	arr[31, 67] = 0
-	arr[31, 68] = 0
-	arr[31, 69] = 0
-	arr[31, 70] = 0
-	arr[31, 71] = 0
-	arr[31, 72] = 0
-	arr[31, 73] = 0
-	arr[31, 74] = 0
-	arr[31, 75] = 0
-	arr[31, 76] = 0
-	arr[31, 77] = 0
-	arr[31, 78] = 0
-	arr[31, 79] = 0
-	arr[31, 80] = 0
-	arr[31, 81] = 0
-	arr[31, 82] = 0
-	arr[31, 83] = 0
-	arr[31, 84] = 0
-	arr[31, 85] = 0
-	arr[31, 86] = -kr[31]
-	arr[31, 87] = kf[31] * y[88]
-	arr[31, 88] = kf[31] * y[87]
-	arr[31, 89] = 0
-	arr[31, 90] = 0
-	arr[31, 91] = 0
-	arr[31, 92] = 0
-	arr[31, 93] = 0
-	arr[31, 94] = 0
-	arr[31, 95] = 0
-	arr[31, 96] = 0
-	arr[31, 97] = 0
-	arr[31, 98] = 0
-	arr[31, 99] = 0
-	arr[31, 100] = 0
-	arr[31, 101] = 0
-	arr[31, 102] = 0
-	arr[31, 103] = 0
-	arr[32, 0] = 0
-	arr[32, 1] = 0
-	arr[32, 2] = 0
-	arr[32, 3] = 0
-	arr[32, 4] = 0
-	arr[32, 5] = 0
-	arr[32, 6] = 0
-	arr[32, 7] = 0
-	arr[32, 8] = 0
-	arr[32, 9] = 0
-	arr[32, 10] = 0
-	arr[32, 11] = 0
-	arr[32, 12] = 0
-	arr[32, 13] = 0
-	arr[32, 14] = 0
-	arr[32, 15] = 0
-	arr[32, 16] = 0
-	arr[32, 17] = 0
-	arr[32, 18] = 0
-	arr[32, 19] = 0
-	arr[32, 20] = 0
-	arr[32, 21] = 0
-	arr[32, 22] = 0
-	arr[32, 23] = 0
-	arr[32, 24] = 0
-	arr[32, 25] = 0
-	arr[32, 26] = 0
-	arr[32, 27] = 0
-	arr[32, 28] = 0
-	arr[32, 29] = 0
-	arr[32, 30] = 0
-	arr[32, 31] = 0
-	arr[32, 32] = 0
-	arr[32, 33] = 0
-	arr[32, 34] = 0
-	arr[32, 35] = 0
-	arr[32, 36] = 0
-	arr[32, 37] = 0
-	arr[32, 38] = 0
-	arr[32, 39] = 0
-	arr[32, 40] = 0
-	arr[32, 41] = 0
-	arr[32, 42] = 0
-	arr[32, 43] = 0
-	arr[32, 44] = 0
-	arr[32, 45] = 0
-	arr[32, 46] = 0
-	arr[32, 47] = 0
-	arr[32, 48] = 0
-	arr[32, 49] = 0
-	arr[32, 50] = 0
-	arr[32, 51] = 0
-	arr[32, 52] = 0
-	arr[32, 53] = 0
-	arr[32, 54] = 0
-	arr[32, 55] = 0
-	arr[32, 56] = 0
-	arr[32, 57] = 0
-	arr[32, 58] = 0
-	arr[32, 59] = 0
-	arr[32, 60] = 0
-	arr[32, 61] = 0
-	arr[32, 62] = 0
-	arr[32, 63] = 0
-	arr[32, 64] = 0
-	arr[32, 65] = 0
-	arr[32, 66] = 0
-	arr[32, 67] = 0
-	arr[32, 68] = 0
-	arr[32, 69] = 0
-	arr[32, 70] = 0
-	arr[32, 71] = 0
-	arr[32, 72] = 0
-	arr[32, 73] = 0
-	arr[32, 74] = 0
-	arr[32, 75] = 0
-	arr[32, 76] = 0
-	arr[32, 77] = 0
-	arr[32, 78] = 0
-	arr[32, 79] = 0
-	arr[32, 80] = 0
-	arr[32, 81] = 0
-	arr[32, 82] = 0
-	arr[32, 83] = 0
-	arr[32, 84] = 0
-	arr[32, 85] = 0
-	arr[32, 86] = 0
-	arr[32, 87] = 0
-	arr[32, 88] = 0
-	arr[32, 89] = -kr[32]
-	arr[32, 90] = kf[32] * y[91]
-	arr[32, 91] = kf[32] * y[90]
-	arr[32, 92] = 0
-	arr[32, 93] = 0
-	arr[32, 94] = 0
-	arr[32, 95] = 0
-	arr[32, 96] = 0
-	arr[32, 97] = 0
-	arr[32, 98] = 0
-	arr[32, 99] = 0
-	arr[32, 100] = 0
-	arr[32, 101] = 0
-	arr[32, 102] = 0
-	arr[32, 103] = 0
-	arr[33, 0] = 0
-	arr[33, 1] = 0
-	arr[33, 2] = 0
-	arr[33, 3] = 0
-	arr[33, 4] = 0
-	arr[33, 5] = 0
-	arr[33, 6] = 0
-	arr[33, 7] = 0
-	arr[33, 8] = 0
-	arr[33, 9] = 0
-	arr[33, 10] = 0
-	arr[33, 11] = 0
-	arr[33, 12] = 0
-	arr[33, 13] = 0
-	arr[33, 14] = 0
-	arr[33, 15] = 0
-	arr[33, 16] = 0
-	arr[33, 17] = 0
-	arr[33, 18] = 0
-	arr[33, 19] = 0
-	arr[33, 20] = 0
-	arr[33, 21] = 0
-	arr[33, 22] = 0
-	arr[33, 23] = 0
-	arr[33, 24] = 0
-	arr[33, 25] = 0
-	arr[33, 26] = 0
-	arr[33, 27] = 0
-	arr[33, 28] = 0
-	arr[33, 29] = 0
-	arr[33, 30] = 0
-	arr[33, 31] = 0
-	arr[33, 32] = 0
-	arr[33, 33] = 0
-	arr[33, 34] = 0
-	arr[33, 35] = 0
-	arr[33, 36] = 0
-	arr[33, 37] = 0
-	arr[33, 38] = 0
-	arr[33, 39] = 0
-	arr[33, 40] = 0
-	arr[33, 41] = 0
-	arr[33, 42] = 0
-	arr[33, 43] = 0
-	arr[33, 44] = 0
-	arr[33, 45] = 0
-	arr[33, 46] = 0
-	arr[33, 47] = 0
-	arr[33, 48] = 0
-	arr[33, 49] = 0
-	arr[33, 50] = 0
-	arr[33, 51] = 0
-	arr[33, 52] = 0
-	arr[33, 53] = 0
-	arr[33, 54] = 0
-	arr[33, 55] = 0
-	arr[33, 56] = 0
-	arr[33, 57] = 0
-	arr[33, 58] = 0
-	arr[33, 59] = 0
-	arr[33, 60] = 0
-	arr[33, 61] = 0
-	arr[33, 62] = 0
-	arr[33, 63] = 0
-	arr[33, 64] = 0
-	arr[33, 65] = 0
-	arr[33, 66] = 0
-	arr[33, 67] = 0
-	arr[33, 68] = 0
-	arr[33, 69] = 0
-	arr[33, 70] = 0
-	arr[33, 71] = 0
-	arr[33, 72] = 0
-	arr[33, 73] = 0
-	arr[33, 74] = 0
-	arr[33, 75] = 0
-	arr[33, 76] = 0
-	arr[33, 77] = 0
-	arr[33, 78] = 0
-	arr[33, 79] = 0
-	arr[33, 80] = 0
-	arr[33, 81] = 0
-	arr[33, 82] = 0
-	arr[33, 83] = 0
-	arr[33, 84] = 0
-	arr[33, 85] = 0
-	arr[33, 86] = 0
-	arr[33, 87] = 0
-	arr[33, 88] = 0
-	arr[33, 89] = 0
-	arr[33, 90] = 0
-	arr[33, 91] = 0
-	arr[33, 92] = -kr[33]
-	arr[33, 93] = kf[33] * y[94]
-	arr[33, 94] = kf[33] * y[93]
-	arr[33, 95] = 0
-	arr[33, 96] = 0
-	arr[33, 97] = 0
-	arr[33, 98] = 0
-	arr[33, 99] = 0
-	arr[33, 100] = 0
-	arr[33, 101] = 0
-	arr[33, 102] = 0
-	arr[33, 103] = 0
-	arr[34, 0] = 0
-	arr[34, 1] = 0
-	arr[34, 2] = 0
-	arr[34, 3] = 0
-	arr[34, 4] = 0
-	arr[34, 5] = 0
-	arr[34, 6] = 0
-	arr[34, 7] = 0
-	arr[34, 8] = 0
-	arr[34, 9] = 0
-	arr[34, 10] = 0
-	arr[34, 11] = 0
-	arr[34, 12] = 0
-	arr[34, 13] = 0
-	arr[34, 14] = 0
-	arr[34, 15] = 0
-	arr[34, 16] = 0
-	arr[34, 17] = 0
-	arr[34, 18] = 0
-	arr[34, 19] = 0
-	arr[34, 20] = 0
-	arr[34, 21] = 0
-	arr[34, 22] = 0
-	arr[34, 23] = 0
-	arr[34, 24] = 0
-	arr[34, 25] = 0
-	arr[34, 26] = 0
-	arr[34, 27] = 0
-	arr[34, 28] = 0
-	arr[34, 29] = 0
-	arr[34, 30] = 0
-	arr[34, 31] = 0
-	arr[34, 32] = 0
-	arr[34, 33] = 0
-	arr[34, 34] = 0
-	arr[34, 35] = 0
-	arr[34, 36] = 0
-	arr[34, 37] = 0
-	arr[34, 38] = 0
-	arr[34, 39] = 0
-	arr[34, 40] = 0
-	arr[34, 41] = 0
-	arr[34, 42] = 0
-	arr[34, 43] = 0
-	arr[34, 44] = 0
-	arr[34, 45] = 0
-	arr[34, 46] = 0
-	arr[34, 47] = 0
-	arr[34, 48] = 0
-	arr[34, 49] = 0
-	arr[34, 50] = 0
-	arr[34, 51] = 0
-	arr[34, 52] = 0
-	arr[34, 53] = 0
-	arr[34, 54] = 0
-	arr[34, 55] = 0
-	arr[34, 56] = 0
-	arr[34, 57] = 0
-	arr[34, 58] = 0
-	arr[34, 59] = 0
-	arr[34, 60] = 0
-	arr[34, 61] = 0
-	arr[34, 62] = 0
-	arr[34, 63] = 0
-	arr[34, 64] = 0
-	arr[34, 65] = 0
-	arr[34, 66] = 0
-	arr[34, 67] = 0
-	arr[34, 68] = 0
-	arr[34, 69] = 0
-	arr[34, 70] = 0
-	arr[34, 71] = 0
-	arr[34, 72] = 0
-	arr[34, 73] = 0
-	arr[34, 74] = 0
-	arr[34, 75] = 0
-	arr[34, 76] = 0
-	arr[34, 77] = 0
-	arr[34, 78] = 0
-	arr[34, 79] = 0
-	arr[34, 80] = 0
-	arr[34, 81] = 0
-	arr[34, 82] = 0
-	arr[34, 83] = 0
-	arr[34, 84] = 0
-	arr[34, 85] = 0
-	arr[34, 86] = 0
-	arr[34, 87] = 0
-	arr[34, 88] = 0
-	arr[34, 89] = 0
-	arr[34, 90] = 0
-	arr[34, 91] = 0
-	arr[34, 92] = 0
-	arr[34, 93] = 0
-	arr[34, 94] = 0
-	arr[34, 95] = -kr[34]
-	arr[34, 96] = kf[34] * y[97]
-	arr[34, 97] = kf[34] * y[96]
-	arr[34, 98] = 0
-	arr[34, 99] = 0
-	arr[34, 100] = 0
-	arr[34, 101] = 0
-	arr[34, 102] = 0
-	arr[34, 103] = 0
-	arr[35, 0] = 0
-	arr[35, 1] = 0
-	arr[35, 2] = 0
-	arr[35, 3] = 0
-	arr[35, 4] = 0
-	arr[35, 5] = 0
-	arr[35, 6] = 0
-	arr[35, 7] = 0
-	arr[35, 8] = 0
-	arr[35, 9] = 0
-	arr[35, 10] = 0
-	arr[35, 11] = 0
-	arr[35, 12] = 0
-	arr[35, 13] = 0
-	arr[35, 14] = 0
-	arr[35, 15] = 0
-	arr[35, 16] = 0
-	arr[35, 17] = 0
-	arr[35, 18] = 0
-	arr[35, 19] = 0
-	arr[35, 20] = 0
-	arr[35, 21] = 0
-	arr[35, 22] = 0
-	arr[35, 23] = 0
-	arr[35, 24] = 0
-	arr[35, 25] = 0
-	arr[35, 26] = 0
-	arr[35, 27] = 0
-	arr[35, 28] = 0
-	arr[35, 29] = 0
-	arr[35, 30] = 0
-	arr[35, 31] = 0
-	arr[35, 32] = 0
-	arr[35, 33] = 0
-	arr[35, 34] = 0
-	arr[35, 35] = 0
-	arr[35, 36] = 0
-	arr[35, 37] = 0
-	arr[35, 38] = 0
-	arr[35, 39] = 0
-	arr[35, 40] = 0
-	arr[35, 41] = 0
-	arr[35, 42] = 0
-	arr[35, 43] = 0
-	arr[35, 44] = 0
-	arr[35, 45] = 0
-	arr[35, 46] = 0
-	arr[35, 47] = 0
-	arr[35, 48] = 0
-	arr[35, 49] = 0
-	arr[35, 50] = 0
-	arr[35, 51] = 0
-	arr[35, 52] = 0
-	arr[35, 53] = 0
-	arr[35, 54] = 0
-	arr[35, 55] = 0
-	arr[35, 56] = 0
-	arr[35, 57] = 0
-	arr[35, 58] = 0
-	arr[35, 59] = 0
-	arr[35, 60] = 0
-	arr[35, 61] = 0
-	arr[35, 62] = 0
-	arr[35, 63] = 0
-	arr[35, 64] = 0
-	arr[35, 65] = 0
-	arr[35, 66] = 0
-	arr[35, 67] = 0
-	arr[35, 68] = 0
-	arr[35, 69] = 0
-	arr[35, 70] = 0
-	arr[35, 71] = 0
-	arr[35, 72] = 0
-	arr[35, 73] = 0
-	arr[35, 74] = 0
-	arr[35, 75] = 0
-	arr[35, 76] = 0
-	arr[35, 77] = 0
-	arr[35, 78] = 0
-	arr[35, 79] = 0
-	arr[35, 80] = 0
-	arr[35, 81] = 0
-	arr[35, 82] = 0
-	arr[35, 83] = 0
-	arr[35, 84] = 0
-	arr[35, 85] = 0
-	arr[35, 86] = 0
-	arr[35, 87] = 0
-	arr[35, 88] = 0
-	arr[35, 89] = 0
-	arr[35, 90] = 0
-	arr[35, 91] = 0
-	arr[35, 92] = 0
-	arr[35, 93] = 0
-	arr[35, 94] = 0
-	arr[35, 95] = 0
-	arr[35, 96] = 0
-	arr[35, 97] = 0
-	arr[35, 98] = -kr[35]
-	arr[35, 99] = kf[35] * y[100]
-	arr[35, 100] = kf[35] * y[99]
-	arr[35, 101] = 0
-	arr[35, 102] = 0
-	arr[35, 103] = 0
-	arr[36, 0] = 0
-	arr[36, 1] = 0
-	arr[36, 2] = 0
-	arr[36, 3] = 0
-	arr[36, 4] = 0
-	arr[36, 5] = 0
-	arr[36, 6] = 0
-	arr[36, 7] = 0
-	arr[36, 8] = 0
-	arr[36, 9] = 0
-	arr[36, 10] = 0
-	arr[36, 11] = 0
-	arr[36, 12] = 0
-	arr[36, 13] = 0
-	arr[36, 14] = 0
-	arr[36, 15] = 0
-	arr[36, 16] = 0
-	arr[36, 17] = 0
-	arr[36, 18] = 0
-	arr[36, 19] = 0
-	arr[36, 20] = 0
-	arr[36, 21] = 0
-	arr[36, 22] = 0
-	arr[36, 23] = 0
-	arr[36, 24] = 0
-	arr[36, 25] = 0
-	arr[36, 26] = 0
-	arr[36, 27] = 0
-	arr[36, 28] = 0
-	arr[36, 29] = 0
-	arr[36, 30] = 0
-	arr[36, 31] = 0
-	arr[36, 32] = 0
-	arr[36, 33] = 0
-	arr[36, 34] = 0
-	arr[36, 35] = 0
-	arr[36, 36] = 0
-	arr[36, 37] = 0
-	arr[36, 38] = 0
-	arr[36, 39] = 0
-	arr[36, 40] = 0
-	arr[36, 41] = 0
-	arr[36, 42] = 0
-	arr[36, 43] = 0
-	arr[36, 44] = 0
-	arr[36, 45] = 0
-	arr[36, 46] = 0
-	arr[36, 47] = 0
-	arr[36, 48] = 0
-	arr[36, 49] = 0
-	arr[36, 50] = 0
-	arr[36, 51] = 0
-	arr[36, 52] = 0
-	arr[36, 53] = 0
-	arr[36, 54] = 0
-	arr[36, 55] = 0
-	arr[36, 56] = 0
-	arr[36, 57] = 0
-	arr[36, 58] = 0
-	arr[36, 59] = 0
-	arr[36, 60] = 0
-	arr[36, 61] = 0
-	arr[36, 62] = 0
-	arr[36, 63] = 0
-	arr[36, 64] = 0
-	arr[36, 65] = 0
-	arr[36, 66] = 0
-	arr[36, 67] = 0
-	arr[36, 68] = 0
-	arr[36, 69] = 0
-	arr[36, 70] = 0
-	arr[36, 71] = 0
-	arr[36, 72] = 0
-	arr[36, 73] = 0
-	arr[36, 74] = 0
-	arr[36, 75] = 0
-	arr[36, 76] = 0
-	arr[36, 77] = 0
-	arr[36, 78] = 0
-	arr[36, 79] = 0
-	arr[36, 80] = 0
-	arr[36, 81] = 0
-	arr[36, 82] = 0
-	arr[36, 83] = 0
-	arr[36, 84] = 0
-	arr[36, 85] = 0
-	arr[36, 86] = 0
-	arr[36, 87] = 0
-	arr[36, 88] = 0
-	arr[36, 89] = 0
-	arr[36, 90] = 0
-	arr[36, 91] = 0
-	arr[36, 92] = 0
-	arr[36, 93] = 0
-	arr[36, 94] = 0
-	arr[36, 95] = 0
-	arr[36, 96] = 0
-	arr[36, 97] = 0
-	arr[36, 98] = 0
-	arr[36, 99] = 0
-	arr[36, 100] = 0
-	arr[36, 101] = -kr[36]
-	arr[36, 102] = kf[36] * y[103]
-	arr[36, 103] = kf[36] * y[102]
-	return arr.reshape(-1)
-
