+++ src/time_numba.py
@@ -4,9 +4,6 @@ import timeit
 from numba import njit
 import numpy as np
 
-from refactored import tcs_rates, tcs_rates_jac, eq_rates, eq_rates_jac
-from refactored_sparse import tcs_rates_sp, tcs_rates_jac_sp, eq_rates_sp, eq_rates_jac_sp
-
 NUMBER = 10_000  # number of times for timeit to run the timed code
 
 # From WCM TwoComponentSystem.
@@ -304,7 +301,7 @@ def builder(dims, *entries):
     return a.reshape(-1)
 
 
-def time_jit(function_name, function_code):
+def time_jit(function_name, function_code, execute=False):
     """Time how long it takes to run the function code, to JIT compile it, and
     to run the JIT compiled code.
     """
@@ -313,7 +310,11 @@ def time_jit(function_name, function_code):
     kf = np.arange(100.0) * 2
     kr = np.arange(100.0) * 10
 
-    if isinstance(function_code, str):
+    if execute:
+        ldict = {}
+        exec(function_code, globals(), ldict)
+        f = ldict[function_name]
+    elif isinstance(function_code, str):
         f = eval(function_code, {'np': np, 'builder': builder}, {})
     else:
         f = function_code
@@ -357,21 +358,66 @@ def refactor_to_lambda(func_string, arr_dims, sparse=False):
     return f_template
 
 
+def refactor_func_string(func_string, func_name, arr_dims, sparse=False):
+    array_string = func_string[36:-19]
+    arr_vals = array_string.split(',')
+    f_template = f"def {func_name}(t, y, kf, kr):\n\tarr = np.zeros({arr_dims})\n"
+    row_idx = -1
+    col_idx = 0
+    for arr_val in arr_vals:
+        col_idx += 1
+        clean_val = arr_val.strip()
+        if clean_val.startswith('['):
+            row_idx += 1
+            col_idx = 0
+            clean_val = clean_val[1:]
+        if clean_val.endswith(']]'):
+            clean_val = clean_val[:-1]
+        if clean_val.endswith(']]'):
+            clean_val = clean_val[:-1]
+        if clean_val == '0]':
+            clean_val = clean_val[:-1]
+        if clean_val == '0' and sparse:
+            continue
+        f_template += f'\tarr[{row_idx}, {col_idx}] = {clean_val}\n'
+    f_template += '\treturn arr.reshape(-1)\n\n'
+    return f_template
+
+
 def time_symbolic_rates():
+    time_jit('tcs_rates_sp', 
+        refactor_func_string(
+            TCS_RATES, 
+            'tcs_rates_sp', 
+            (29, 1), 
+            sparse=True), 
+        execute=True)
+    time_jit('tcs_rates_jac_sp', 
+        refactor_func_string(
+            TCS_RATES_JACOBIAN, 
+            'tcs_rates_jac_sp', 
+            (29, 41), 
+            sparse=True), 
+        execute=True)
+    time_jit('eq_rates_sp', 
+        refactor_func_string(
+            EQUILIBRIUM_RATES, 
+            'eq_rates_sp', 
+            (37, 1), 
+            sparse=True), 
+        execute=True)
+    time_jit('eq_rates_jac_sp', 
+        refactor_func_string(
+            EQUILIBRIUM_RATES_JACOBIAN, 
+            'eq_rates_jac_sp', 
+            (37, 104), 
+            sparse=True), 
+        execute=True)
+    
     time_jit('tcs_rates_sp_lambda', refactor_to_lambda(TCS_RATES, (29, 1), True))
     time_jit('tcs_rates_jac_sp_lambda', refactor_to_lambda(TCS_RATES_JACOBIAN, (29, 41), True))
     time_jit('eq_rates_sp_lambda', refactor_to_lambda(EQUILIBRIUM_RATES, (37, 1), True))
     time_jit('eq_rates_jac_sp_lambda', refactor_to_lambda(EQUILIBRIUM_RATES_JACOBIAN, (37, 104), True))
-    
-    time_jit('tcs_rates_sp', tcs_rates_sp)
-    time_jit('tcs_rates_jac_sp', tcs_rates_jac_sp)
-    time_jit('eq_rates_sp', eq_rates_sp)
-    time_jit('eq_rates_jac_sp', eq_rates_jac_sp)
-
-    time_jit('tcs_rates', tcs_rates)
-    time_jit('tcs_rates_jac', tcs_rates_jac)
-    time_jit('eq_rates', eq_rates)
-    time_jit('eq_rates_jac', eq_rates_jac)
 
     time_jit('2CS RATES', TCS_RATES)
     time_jit('2CS RATES JACOBIAN', TCS_RATES_JACOBIAN)
